[#253] Decompose jeod_runner/src/lib.rs into a simulation/ submodule
3,276-line `crates/jeod_runner/src/lib.rs` carved into 7 focused
files; `DetachedSubtreeState` relocated to `jeod_dynamics`; the
Phase-10 `pub use jeod_sim::{...}` re-export block dropped and ~30
consumers retargeted to import from `jeod_sim` directly.

The whole change is mechanical: methods/types move between files,
re-exports update, no behavior change. Tier 3 baselines unchanged
(78/78 matched), trajectory residuals identical to pre-refactor.

## Layout (Task A — main file split)

`crates/jeod_runner/src/lib.rs` shrinks from 3,276 → 61 lines —
crate-level docs and `pub use` re-exports only. The `Simulation`
struct definition + 43 public methods live in:

```
crates/jeod_runner/src/simulation/
├── mod.rs        ~210 L  // Simulation struct, new(), set_dt,
│                          // elapsed, num_bodies
├── types.rs      ~270 L  // SimBody, VehicleOutput, ContactPairConfig,
│                          // SourceFrameIds, GravityData
├── sources.rs    ~260 L  // gravity-source registration / accessors
├── bodies.rs     ~410 L  // body lifecycle, accessors, setters,
│                          // contact-pair registration,
│                          // subtree_composite_inertial
├── mass_tree.rs  ~600 L  // attach/detach machinery
│                          // (detach_subtree, attach_subtree_aligned)
├── validate.rs   ~340 L  // validate() — pre-step setup-time validation
└── step.rs     ~1200 L  // step / step_internal / step_n / step_until
```

Rust's standard "split impl across child modules" pattern: child
files reopen `impl Simulation { ... }` and retain access to the
struct's private fields/methods because the children inherit
visibility from the parent module.

## Task C — relocate `DetachedSubtreeState` + `step_ballistic`

Pure rigid-body kinematics with zero `bevy_jeod` or `jeod_sim`
consumers. Moved to `crates/jeod_dynamics/src/subtree.rs`,
adjacent to the `advance_left_quat_body_rate` helper that
`step_ballistic` already delegates to.
`crates/jeod_runner/src/lib.rs` keeps a `pub use
jeod_dynamics::DetachedSubtreeState;` for API stability.

## Task D — Phase-10 re-export cleanup

The 11-line `pub use jeod_sim::{...}` block at the previous
`lib.rs:64-74` (12 names: `VehicleConfig`, `GravitySourceEntry`,
`SrpModel`, `FrameSwitchConfig`, `SwitchSense`, ... ) was
self-tagged for removal in the in-source comment. Dropped. ~30
consumer files (workspace tests + examples + a few crate-internal
sites) updated to import the relocated names from `jeod_sim`
directly via separate `use jeod_sim::{...};` lines. The
crate-wide `pub use jeod_sim;` re-export remains, so straggling
`jeod_runner::jeod_sim::VehicleConfig`-style paths still work.

## Architectural placement check (per user prompt)

Verified before splitting that nothing in `jeod_runner/src/lib.rs`
belongs in `jeod_sim` (the API surface for the Bevy adapter):
`Simulation` and its methods are runner-specific (Bevy uses
its own systems calling `jeod_sim` functions); `VehicleOutput` is
explicitly tagged "runner-side" in `jeod_sim`'s docs;
`ContactPairConfig` is registration (Bevy uses ECS pair
registration, `jeod_sim` already owns evaluation).
`DetachedSubtreeState` was the one cross-crate move; it went to
`jeod_dynamics` (foundational physics), not `jeod_sim`
(orchestration).

## Deferred

Task B from the plan — splitting `step_internal` per JEOD
integration stage (`simulation/step/{ephemeris,environment,
interactions,integrate,derived}.rs`) — is **not** in this commit.
The state-flow refactor it requires (threading `body_states_pre`,
per-source gravity accels, atmospheric density, etc. through stage
methods via tuple returns or a `StepScratch<'a>` struct) goes
beyond a mechanical split and would need a separate, careful pass
to preserve bit-equality on Tier 3 baselines. `step.rs` stays as
one ~1,200-line file with `step_internal` intact. Stage banners
already inside the function (`// ── 1. Time update ──`, etc.) keep
navigation manageable, and #253 explicitly listed this stretch
goal as a candidate follow-up.

## Verification

- `cargo build --workspace`: clean
- `cargo fmt --check`: clean
- `cargo clippy --workspace --tests -- -D warnings`: clean
- `RUSTDOCFLAGS=-D warnings cargo doc --workspace --no-deps`: clean
- `cargo nextest run --workspace -E 'not test(tier3_)'`: 823/823 pass
- `cargo nextest run --workspace -E 'test(tier3_) and not test(earth_moon)'`: 308/308 pass
- `cargo run -p jeod_test_data --bin tier3_baseline_diff`: OK (78 matched, 0 missing, 0 new)

Closes #253 (file split + DetachedSubtreeState relocation +
Phase-10 cleanup); per-stage `step_internal` decomposition tracked
as the remaining stretch goal.
