+++ crates/jeod_runner/src/lib.rs
@@ -21,31 +21,13 @@
 #![forbid(unsafe_code)]
 #![warn(missing_docs)]
 
-use std::collections::HashMap;
-
-use glam::{DMat3, DVec3};
-
-use jeod_dynamics::{combine_states_at_attach, AttachCombineInputs, MassBodyId, MassPointState};
-use jeod_frames::{FrameTree, RefFrameKind, RefFrameRot, RefFrameState, RefFrameTrans};
-use jeod_sim::atmosphere::{evaluate_atmosphere, AtmosphereConfig};
-use jeod_sim::forces::collect_and_resolve_forces;
-use jeod_sim::gravity::accumulate_gravity;
-use jeod_sim::integration::{integrate_bodies_contact_coupled, integrate_body, CoupledBodyInput};
-use jeod_sim::validation::ValidationError;
-use jeod_sim::{
-    evaluate_contact_pair, integrate_body_coupled, AerodynamicForce, AtmosphereState, ContactFacet,
-    CoupledStageEval, DragConfig, DynamicsConfig, EulerSequence, FrameDerivatives, GeodeticState,
-    GravityAcceleration, GravityControls, GravitySource, JeodQuat, LvlhFrame, MassProperties,
-    OrbitalElements, RadiationForce, RotationalState, SimulationTime, TotalForce,
-    TranslationalState,
-};
-
 pub mod branded;
 pub mod builder;
 pub mod error;
 pub mod prelude;
 #[cfg(feature = "verification")]
 pub mod run_verification;
+mod simulation;
 
 pub use branded::{BodyIdx, BrandedSimulation, SourceIdx};
 pub use error::StepError;
@@ -61,3216 +43,17 @@ pub use builder::SimulationBuilderExt;
 #[cfg(feature = "verification")]
 pub use run_verification::VerificationCaseExt;
 
-// Re-exports of types relocated from jeod_runner to jeod_sim in Phase 6 of
-// #101. External consumers of `jeod_runner::{VehicleConfig,GravitySourceEntry,
-// SrpModel,...}` continue to compile via these re-exports; the canonical
-// definitions now live in `jeod_sim`. Phase 10 will remove the re-exports.
-pub use jeod_sim::{
-    DerivedStateConfig, EarthLightingConfig, FrameSwitchConfig, GeodeticConfig, GravitySourceEntry,
-    MassTreeAttachment, ShadowBody, SimulationBuilder, SrpModel, SwitchSense, VehicleBuilder,
-    VehicleConfig,
-};
+// (Phase-10 cleanup, issue #253) The 12 types relocated from `jeod_runner`
+// to `jeod_sim` in Phase 6 of #101 are no longer re-exported here. Consumers
+// import them directly from `jeod_sim`:
+//   `use jeod_sim::{VehicleConfig, GravitySourceEntry, SrpModel, ...};`
+//
+// `jeod_runner::jeod_sim` is still re-exported above so any straggling
+// `jeod_runner::jeod_sim::VehicleConfig`-style paths continue to compile.
 
 // Re-export FrameId for downstream API.
 pub use jeod_frames::FrameId;
 
-/// Composite-body inertial state of a free-flying mass-tree subtree
-/// (i.e. a tree root that is not the integrated body). Populated when
-/// the subtree is detached from the integrated body's tree, propagated
-/// each step (ballistically — tree-only subtrees have no gravity
-/// controls of their own), consumed when the subtree is re-attached.
-///
-/// All fields are in the simulation's integration frame (typically
-/// Earth.inertial). Stored on [`Simulation::detached_subtrees`].
-#[derive(Debug, Clone, Copy)]
-pub struct DetachedSubtreeState {
-    /// Inertial position of the subtree's composite CoM.
-    pub composite_position: DVec3,
-    /// Inertial velocity of the subtree's composite CoM.
-    pub composite_velocity: DVec3,
-    /// Inertial-to-body rotation (`q_parent_this`, JEOD scalar-first
-    /// left-quat convention — same orientation semantics as
-    /// `RotationalState::quaternion`).
-    pub composite_attitude: JeodQuat,
-    /// Angular velocity in the subtree's body frame.
-    pub composite_ang_vel_body: DVec3,
-}
-
-impl DetachedSubtreeState {
-    /// Convert to a [`RefFrameState`] for use with the propagation
-    /// helpers in `jeod_dynamics`.
-    pub fn to_ref_frame_state(&self) -> RefFrameState {
-        RefFrameState {
-            trans: RefFrameTrans {
-                position: self.composite_position,
-                velocity: self.composite_velocity,
-            },
-            rot: RefFrameRot {
-                q_parent_this: self.composite_attitude,
-                t_parent_this: self.composite_attitude.left_quat_to_transformation(),
-                ang_vel_this: self.composite_ang_vel_body,
-            },
-        }
-    }
-
-    /// Construct from a [`RefFrameState`].
-    pub fn from_ref_frame_state(state: &RefFrameState) -> Self {
-        Self {
-            composite_position: state.trans.position,
-            composite_velocity: state.trans.velocity,
-            composite_attitude: state.rot.q_parent_this,
-            composite_ang_vel_body: state.rot.ang_vel_this,
-        }
-    }
-
-    /// Advance the subtree ballistically by `dt` seconds (no force, no
-    /// torque). Position drifts at `composite_velocity`; attitude
-    /// rotates at `composite_ang_vel_body` via
-    /// [`jeod_dynamics::advance_left_quat_body_rate`] (which owns the
-    /// JEOD left-multiply convention so this site can't get it wrong);
-    /// velocity and ang_vel are unchanged.
-    pub fn step_ballistic(&mut self, dt: f64) {
-        self.composite_position += self.composite_velocity * dt;
-        self.composite_attitude = jeod_dynamics::advance_left_quat_body_rate(
-            self.composite_attitude,
-            self.composite_ang_vel_body,
-            dt,
-        );
-    }
-}
-
-// ══════════════════════════════════════════════════════════════════════════════
-// Contact pair
-// ══════════════════════════════════════════════════════════════════════════════
-
-/// Registration of a contact interaction between two bodies.
-///
-/// Port of JEOD's `ContactPair` — the two facets are registered with the
-/// `Contact` manager, which runs `check_contact()` at each derivative
-/// evaluation (see `contact.sm`). In this runner, registered pairs are
-/// evaluated at every RK4 stage inside [`Simulation::step`] when any pairs
-/// are present.
-#[derive(Debug, Clone)]
-pub struct ContactPairConfig {
-    /// Index of body A (the "subject" in JEOD terminology).
-    pub body_a: usize,
-    /// Facet on body A (shape positions in A's structural frame).
-    pub facet_a: ContactFacet,
-    /// Index of body B (the "target" in JEOD terminology).
-    pub body_b: usize,
-    /// Facet on body B (shape positions in B's structural frame).
-    pub facet_b: ContactFacet,
-}
-
-// ══════════════════════════════════════════════════════════════════════════════
-// Frame tree source registry
-// ══════════════════════════════════════════════════════════════════════════════
-
-/// Maps a gravity source to its frame tree nodes.
-struct SourceFrameIds {
-    /// Inertial frame for this source (e.g., "Earth.inertial").
-    inertial: FrameId,
-    /// Planet-fixed frame (e.g., "Earth.pfix"), if the source has a rotation model.
-    pfix: Option<FrameId>,
-}
-
-/// Gravity-specific data associated with a source (decoupled from frame tree).
-///
-/// The frame tree stores position/velocity/rotation state; this struct stores
-/// the physical gravity model data that lives alongside it. The `velocity`
-/// field stores source velocity for relativistic corrections — for central
-/// bodies at the root frame, the tree node has zero velocity but the source
-/// may still have physical velocity (e.g., Sun orbiting the barycenter).
-struct GravityData {
-    /// Physical gravity source (mu, model: PointMass or SphericalHarmonics).
-    source: GravitySource,
-    /// Source velocity in the inertial frame (m/s). Used for relativistic
-    /// corrections. Stored here rather than in the tree because the root
-    /// frame's velocity must be zero (it's the reference origin).
-    velocity: DVec3,
-    /// Tidal ΔC20 to add to the base C20 coefficient. Updated each step.
-    delta_c20: f64,
-    /// Tidal configuration. When `Some`, the simulation computes ΔC20 each step.
-    tidal_config: Option<jeod_gravity::tides::TidalConfig>,
-    /// Rotation model for updating planet-fixed frame each step.
-    rotation_model: RotationModel,
-    /// Sidereal angular velocity (rad/s) for the planet-fixed frame's
-    /// `ang_vel_this`. Sourced from `PlanetConfig::omega` at setup time.
-    /// JEOD sets this as `[0, 0, planet_omega]` in `planet_rnp.cc`.
-    planet_omega: f64,
-}
-
-// ══════════════════════════════════════════════════════════════════════════════
-// Vehicle output (public, read-only view of results after step)
-// ══════════════════════════════════════════════════════════════════════════════
-
-/// Read-only view of vehicle state after stepping.
-///
-/// Returned by [`Simulation::body`]. Contains the current integrated state
-/// plus any derived states that were configured.
-#[derive(Debug, Clone)]
-pub struct VehicleOutput {
-    /// Current translational state (position, velocity) in the integration frame.
-    pub trans: TranslationalState,
-    /// Frame ID of the current integration frame in the simulation's frame tree.
-    pub integ_frame_id: FrameId,
-    /// Current rotational state (quaternion, angular velocity). `None` for 3-DOF.
-    pub rot: Option<RotationalState>,
-    /// Total translational acceleration in the integration frame (m/s²) at the
-    /// end of the last `step()`. Sum of gravity and non-gravity contributions —
-    /// mirrors JEOD's `derivs.trans_accel`. Zero before the first `step()`.
-    pub trans_accel: DVec3,
-    /// Total rotational acceleration in the body frame (rad/s²) at the end of
-    /// the last `step()` — mirrors JEOD's `derivs.rot_accel`. `None` for 3-DOF
-    /// bodies; zero before the first `step()`.
-    pub rot_accel: Option<DVec3>,
-    /// Orbital elements from the latest step.
-    pub orbital_elements: Option<OrbitalElements>,
-    /// Euler angles `[phi, theta, psi]` from the latest step.
-    pub euler_angles: Option<[f64; 3]>,
-    /// LVLH frame from the latest step.
-    pub lvlh_frame: Option<LvlhFrame>,
-    /// Geodetic state (latitude, longitude, altitude).
-    pub geodetic_state: Option<GeodeticState>,
-    /// Solar beta angle (radians).
-    pub solar_beta: Option<f64>,
-    /// Earth lighting state (sun/moon occlusion, albedo).
-    pub earth_lighting: Option<jeod_interactions::earth_lighting::EarthLightingState>,
-}
-
-// ══════════════════════════════════════════════════════════════════════════════
-// Internal body state (private — not part of public API)
-// ══════════════════════════════════════════════════════════════════════════════
-
-/// Internal per-body simulation state. Combines user config with bookkeeping
-/// and output fields. Not exposed publicly — users interact through
-/// [`VehicleConfig`] (input) and [`VehicleOutput`] (output).
-struct SimBody {
-    // ── Config (from VehicleConfig) ──
-    trans: TranslationalState,
-    rot: Option<RotationalState>,
-    mass: Option<MassProperties>,
-    /// If this body participates in a mass tree, its node ID.
-    mass_body_id: Option<jeod_dynamics::MassBodyId>,
-    config: DynamicsConfig,
-    gravity_controls: GravityControls<usize>,
-    integrator: jeod_dynamics::IntegratorType,
-    drag: Option<DragConfig>,
-    flat_plate_state: Option<jeod_sim::FlatPlateState>,
-    cannonball_srp: Option<(f64, f64, f64)>,
-    shadow_body: Option<(usize, f64)>,
-    t_struct_body: DMat3,
-    compute_gravity_torque: bool,
-    atmospheric_state: Option<AtmosphereState>,
-    external_force: DVec3,
-    external_torque: DVec3,
-
-    // ── Frame switching ──
-    integ_frame_id: FrameId,
-    body_frame_id: FrameId,
-    frame_switches: Vec<FrameSwitchConfig>,
-
-    // ── Bookkeeping (written each step, not user-visible) ──
-    gravity_accel: GravityAcceleration,
-    total_force: TotalForce,
-    frame_derivs: FrameDerivatives,
-    aero_force: Option<AerodynamicForce>,
-    radiation_force: Option<RadiationForce>,
-    gravity_torque: Option<DVec3>,
-
-    // ── Derived state config ──
-    orbital_elements_source: Option<usize>,
-    euler_sequence: Option<EulerSequence>,
-    compute_lvlh: bool,
-    geodetic_planet: Option<(usize, f64, f64)>,
-    compute_solar_beta: bool,
-    earth_lighting_config: Option<(f64, f64, f64)>,
-
-    // ── Derived state outputs ──
-    orbital_elements: Option<OrbitalElements>,
-    euler_angles: Option<[f64; 3]>,
-    lvlh_frame: Option<LvlhFrame>,
-    geodetic_state: Option<GeodeticState>,
-    solar_beta: Option<f64>,
-    earth_lighting: Option<jeod_interactions::earth_lighting::EarthLightingState>,
-
-    // ── Integrator state ──
-    gj_state: Option<jeod_dynamics::GaussJacksonState>,
-    abm4_state: Option<jeod_dynamics::Abm4State>,
-}
-
-impl SimBody {
-    /// Convert a user-facing VehicleConfig into an internal SimBody.
-    fn from_config(config: VehicleConfig, integ_frame_id: FrameId, body_frame_id: FrameId) -> Self {
-        let has_rot = config.rot.is_some();
-        let dynamics_config = DynamicsConfig {
-            translational_dynamics: true,
-            rotational_dynamics: has_rot,
-            three_dof: !has_rot,
-        };
-
-        let (flat_plate_state, cannonball_srp) = match config.srp {
-            Some(SrpModel::FlatPlate(fps)) => (Some(fps), None),
-            Some(SrpModel::Cannonball {
-                cx_area,
-                albedo,
-                diffuse,
-            }) => (None, Some((cx_area, albedo, diffuse))),
-            None => (None, None),
-        };
-
-        let shadow_body = config.shadow_body.map(|sb| (sb.source_idx, sb.radius));
-
-        let has_drag = config.drag.is_some();
-        let atmospheric_state = if has_drag {
-            Some(AtmosphereState::default())
-        } else {
-            None
-        };
-
-        Self {
-            trans: config.trans,
-            rot: config.rot,
-            mass: config.mass,
-            mass_body_id: None,
-            config: dynamics_config,
-            gravity_controls: config.gravity_controls,
-            integrator: config.integrator,
-            drag: config.drag,
-            flat_plate_state,
-            cannonball_srp,
-            shadow_body,
-            t_struct_body: config.t_struct_body,
-            compute_gravity_torque: config.compute_gravity_gradient,
-            atmospheric_state,
-            external_force: config.external_force,
-            external_torque: config.external_torque,
-
-            integ_frame_id,
-            body_frame_id,
-            frame_switches: config.frame_switches,
-
-            gravity_accel: GravityAcceleration::default(),
-            total_force: TotalForce::default(),
-            frame_derivs: FrameDerivatives::default(),
-            aero_force: None,
-            radiation_force: None,
-            gravity_torque: None,
-
-            orbital_elements_source: config.derived.orbital_elements_source,
-            euler_sequence: config.derived.euler_sequence,
-            compute_lvlh: config.derived.lvlh,
-            geodetic_planet: config
-                .derived
-                .geodetic
-                .map(|g| (g.source_idx, g.r_eq, g.r_pol)),
-            compute_solar_beta: config.derived.solar_beta,
-            earth_lighting_config: config
-                .derived
-                .earth_lighting
-                .map(|e| (e.earth_radius, e.moon_radius, e.sun_radius)),
-
-            orbital_elements: None,
-            euler_angles: None,
-            lvlh_frame: None,
-            geodetic_state: None,
-            solar_beta: None,
-            earth_lighting: None,
-
-            gj_state: None,
-            abm4_state: None,
-        }
-    }
-
-    /// Create a VehicleOutput view of the current state.
-    fn output(&self) -> VehicleOutput {
-        VehicleOutput {
-            trans: self.trans,
-            integ_frame_id: self.integ_frame_id,
-            rot: self.rot,
-            trans_accel: self.frame_derivs.trans_accel,
-            rot_accel: self.rot.map(|_| self.frame_derivs.rot_accel),
-            orbital_elements: self.orbital_elements.clone(),
-            euler_angles: self.euler_angles,
-            lvlh_frame: self.lvlh_frame,
-            geodetic_state: self.geodetic_state,
-            solar_beta: self.solar_beta,
-            earth_lighting: self.earth_lighting.clone(),
-        }
-    }
-}
-
-// ══════════════════════════════════════════════════════════════════════════════
-// Simulation
-// ══════════════════════════════════════════════════════════════════════════════
-
-/// ECS-agnostic simulation runner.
-///
-/// Owns all simulation state and runs the JEOD pipeline in `step()`.
-/// This is the **non-ECS** path — ECS adapters should call the per-body
-/// functions (`accumulate_gravity`, `evaluate_atmosphere`, etc.) directly
-/// from their system functions.
-///
-/// # Public API conventions
-///
-/// The methods on `Simulation` group into four families:
-///
-/// - **Source registry** (`add_source`, `set_source_*`, `source_*`) — gravity
-///   sources, ephemeris, planet rotation, tides.
-/// - **Body registry** (`add_body`, `body`, `set_body_*`) — dynamic vehicles.
-/// - **Mass tree** (`add_body_to_tree`, `attach`, `detach`,
-///   `sync_body_mass_from_tree`) — multi-body composites.
-/// - **Lifecycle** (`validate`, `step`, `step_n`, `step_until`, `set_dt`,
-///   `elapsed`).
-///
-/// # Error handling
-///
-/// Validation can happen either at construction time or after mutation.
-/// Builder-based constructors that validate configuration
-/// (`Simulation::from_builder` and `SimulationBuilderExt::build`) return
-/// `Result`. After construction, `validate()` is the only `Result`-returning
-/// instance method on `Simulation`; it batches configuration errors before
-/// stepping. Other runtime methods panic on misuse with a method-name-prefixed
-/// message (e.g. `"set_source_position: source index 7 out of range"`).
-/// Out-of-range indices, configuration conflicts, and numerical
-/// preconditions are programmer errors, not runtime conditions.
-///
-/// # Example
-/// ```
-/// use jeod_runner::SimulationBuilderExt;
-/// use jeod_sim::recipes::Mission;
-///
-/// let mut sim = Mission::iss_leo().into_builder().build().unwrap();
-/// sim.step_n(10);
-/// let output = sim.body(0);
-/// assert!(output.trans.position.length() > 6_000_000.0);
-/// ```
-pub struct Simulation {
-    /// Simulation time (TAI, UTC, TDB, GMST, etc.).
-    pub time: SimulationTime,
-    /// Dynamic bodies (internal, private).
-    // JEOD_INV: DS.01 — private to prevent runtime mutation of derived-state config
-    bodies: Vec<SimBody>,
-    /// Reference frame tree — single source of truth for celestial body positions,
-    /// velocities, and rotations. Updated each step from ephemeris data.
-    /// Private to protect invariants; use [`frame_tree()`](Self::frame_tree) for
-    /// read-only access.
-    frame_tree: FrameTree,
-    /// Root inertial frame ID for this simulation. This is the integration-origin
-    /// frame to which all positions are relative, and it is not necessarily
-    /// `Earth.inertial` (for example, it may be renamed to match the configured
-    /// central body, such as `Mars.inertial`).
-    pub root_frame_id: FrameId,
-    /// Per-source frame tree node IDs (parallel to `gravity_data`).
-    source_frame_ids: Vec<SourceFrameIds>,
-    /// Per-source gravity model data (parallel to `source_frame_ids`).
-    gravity_data: Vec<GravityData>,
-    /// Per-source ephemeris body mapping (parallel to `source_frame_ids`).
-    source_ephem_bodies: Vec<Option<(jeod_sim::EphemerisBody, jeod_sim::EphemerisBody)>>,
-    /// Atmosphere configuration. `None` disables atmosphere for all bodies.
-    pub atmosphere: Option<AtmosphereConfig>,
-    /// Source index for the planet whose rotation is used for atmosphere.
-    pub atmosphere_planet_source: Option<usize>,
-    /// Source index for the Sun (used by SRP and earth lighting).
-    pub sun_source: Option<usize>,
-    /// Source index for the Moon (used by earth lighting).
-    pub moon_source: Option<usize>,
-    /// Polar motion parameters (xp, yp) in radians. When `Some`, the RNP
-    /// composition includes polar motion: W(xp,yp) × R(GAST) × N × P.
-    /// When `None`, polar motion is omitted (matches JEOD `enable_polar=false`).
-    pub polar_motion: Option<(f64, f64)>,
-    /// Integration timestep (seconds).
-    pub dt: f64,
-    /// Optional ephemeris for per-step source position updates.
-    pub ephemeris: Option<jeod_sim::Ephemeris>,
-    /// Optional mass tree for multi-body vehicles (attach/detach/staging).
-    /// Bodies participating in the tree have `SimBody::mass_body_id` set.
-    pub mass_tree: Option<jeod_dynamics::MassTree>,
-    /// Composite-body inertial state of free-flying mass-tree subtrees
-    /// that have been detached from the integrated body's tree but not
-    /// yet re-attached. Populated by [`detach_subtree`](Self::detach_subtree),
-    /// propagated each step by [`step_detached_subtrees`](Self::step_detached_subtrees)
-    /// (called from `step`), consumed by
-    /// [`attach_subtree_aligned`](Self::attach_subtree_aligned). Tree-only
-    /// subtrees never become Simulation bodies — only their composite-CoM
-    /// state is tracked, sufficient for JEOD's `attach_child`
-    /// momentum-conservation algorithm to combine them back in.
-    pub detached_subtrees: HashMap<MassBodyId, DetachedSubtreeState>,
-    /// Registered contact pairs (inter-body spring-damper contact).
-    ///
-    /// When non-empty, `step_internal` uses a multi-body coupled RK4 path in
-    /// which contact forces are recomputed at each of the 4 RK4 stages with
-    /// each pair's intermediate states. This matches JEOD's `check_contact()`
-    /// derivative-class job. Only RK4 + 6-DOF is supported; adding a pair
-    /// while a body uses non-RK4 or 3-DOF is a validation error.
-    contact_pairs: Vec<ContactPairConfig>,
-    /// Preallocated scratch buffers for the coupled RK4 integrator. Retained
-    /// across steps so the inner loop is allocation-free once the body count
-    /// stabilizes.
-    coupled_integ_scratch: jeod_sim::integration::CoupledIntegScratch,
-}
-
-impl Simulation {
-    /// Create a new simulation with the given initial time and timestep.
-    ///
-    /// Creates a frame tree whose root is initially named "Earth.inertial"
-    /// and may be renamed when a central source is added via
-    /// [`Self::add_source`].
-    /// All positions are relative to this root frame regardless of its name.
-    pub fn new(time: SimulationTime, dt: f64) -> Self {
-        let mut frame_tree = FrameTree::new();
-        let root_frame_id = frame_tree.add_root("Earth.inertial".into(), RefFrameKind::Inertial);
-        Self {
-            time,
-            bodies: Vec::new(),
-            frame_tree,
-            root_frame_id,
-            source_frame_ids: Vec::new(),
-            gravity_data: Vec::new(),
-            source_ephem_bodies: Vec::new(),
-            atmosphere: None,
-            atmosphere_planet_source: None,
-            sun_source: None,
-            moon_source: None,
-            polar_motion: None,
-            dt,
-            ephemeris: None,
-            mass_tree: None,
-            detached_subtrees: HashMap::new(),
-            contact_pairs: Vec::new(),
-            coupled_integ_scratch: jeod_sim::integration::CoupledIntegScratch::new(),
-        }
-    }
-
-    /// Register a contact interaction between two bodies.
-    ///
-    /// Once registered, contact forces between these facets are evaluated at
-    /// each RK4 stage of [`step`](Self::step). This matches JEOD's
-    /// derivative-class `check_contact()` scheduling in `SIM_contact`.
-    ///
-    /// The force acts on body A (`facet_a`); the equal-and-opposite force
-    /// acts on body B. Torques are accumulated about each body's CoM.
-    ///
-    /// # Panics
-    /// * Either `body_a` or `body_b` is out of range for the registered bodies.
-    /// * `body_a == body_b` — contact pair bodies must be distinct
-    ///   (JEOD_INV: IN.30, matching JEOD's `unique_pair` invariant).
-    /// * `facet_a.material != facet_b.material`. JEOD parks the
-    ///   spring/damper/friction parameters on a single `SpringPairInteraction`
-    ///   per pair, so both facets must carry identical
-    ///   [`ContactMaterial`](jeod_interactions::ContactMaterial) values.
-    ///   Panic here instead of deferring until the first integrator step.
-    pub fn register_contact_pair(
-        &mut self,
-        body_a: usize,
-        facet_a: ContactFacet,
-        body_b: usize,
-        facet_b: ContactFacet,
-    ) {
-        assert!(
-            body_a < self.bodies.len(),
-            "register_contact_pair: body_a index {body_a} out of range ({} bodies)",
-            self.bodies.len()
-        );
-        assert!(
-            body_b < self.bodies.len(),
-            "register_contact_pair: body_b index {body_b} out of range ({} bodies)",
-            self.bodies.len()
-        );
-        // JEOD_INV: IN.30 — contact pair bodies must be distinct (JEOD `unique_pair`)
-        assert_ne!(
-            body_a, body_b,
-            "register_contact_pair: body A and body B must be distinct (got both = {body_a})"
-        );
-        // JEOD pairs a single `SpringPairInteraction` to each facet pair, so
-        // both facets must carry identical material parameters. Enforce here
-        // rather than inside `compute_contact_force` at first step.
-        assert_eq!(
-            facet_a.material, facet_b.material,
-            "register_contact_pair: facet_a.material and facet_b.material must be equal \
-             (JEOD pairs a single SpringPairInteraction to each facet pair)"
-        );
-        self.contact_pairs.push(ContactPairConfig {
-            body_a,
-            facet_a,
-            body_b,
-            facet_b,
-        });
-    }
-
-    /// Number of registered contact pairs.
-    pub fn num_contact_pairs(&self) -> usize {
-        self.contact_pairs.len()
-    }
-
-    /// Add a gravity source. Returns its index for use in `GravityControls`.
-    ///
-    /// Sources with `central: true` (set by [`GravitySourceEntry::central_body`]
-    /// and [`GravitySourceEntry::central_body_sh`]) are mapped to the root frame.
-    /// Non-central sources get child inertial frames under the root.
-    ///
-    /// Only one central source may be added; a second will panic.
-    ///
-    /// If the source has a rotation model, a planet-fixed child frame is also
-    /// created under the source's inertial frame.
-    pub fn add_source(&mut self, name: impl Into<String>, entry: GravitySourceEntry) -> usize {
-        let idx = self.gravity_data.len();
-        let name = name.into();
-
-        // Central bodies map to the root frame; third bodies get child frames.
-        // Only one central source is allowed (the root can't be shared).
-        let inertial_name = format!("{name}.inertial");
-        let inertial_id = if entry.central {
-            assert!(
-                !self
-                    .source_frame_ids
-                    .iter()
-                    .any(|sf| sf.inertial == self.root_frame_id),
-                "add_source: a central source already maps to root_frame_id. \
-                 Only one central source is allowed per simulation."
-            );
-            assert!(
-                entry.position == DVec3::ZERO,
-                "add_source: central sources must have zero position because they map \
-                 directly to root_frame_id."
-            );
-            // Central body: use the root frame directly. Rename to match.
-            // `entry.velocity` is stored in `gravity_data` for relativistic
-            // corrections, but is not applied as root-frame kinematics.
-            self.frame_tree.get_mut(self.root_frame_id).name = inertial_name;
-            self.root_frame_id
-        } else {
-            self.frame_tree.add_child(
-                self.root_frame_id,
-                inertial_name,
-                RefFrameKind::Inertial,
-                RefFrameState {
-                    trans: RefFrameTrans {
-                        position: entry.position,
-                        velocity: entry.velocity,
-                    },
-                    rot: RefFrameRot::default(),
-                },
-            )
-        };
-
-        // Create a planet-fixed child when the source has a rotation model or
-        // an explicit inertial-to-pfix transform. This ensures a fixed initial
-        // orientation is not silently ignored when rotation_model is None.
-        let pfix_id =
-            if entry.rotation_model != RotationModel::None || entry.t_inertial_pfix.is_some() {
-                let pfix_name = format!("{name}.pfix");
-                let rot = if let Some(t) = entry.t_inertial_pfix {
-                    RefFrameRot {
-                        q_parent_this: JeodQuat::left_quat_from_transformation(&t),
-                        t_parent_this: t,
-                        ang_vel_this: DVec3::ZERO,
-                    }
-                } else {
-                    RefFrameRot::default()
-                };
-                Some(self.frame_tree.add_child(
-                    inertial_id,
-                    pfix_name,
-                    RefFrameKind::PlanetFixed,
-                    RefFrameState {
-                        trans: RefFrameTrans::default(),
-                        rot,
-                    },
-                ))
-            } else {
-                None
-            };
-
-        // Tidal ΔC20 requires a planet-fixed frame for the rotation matrix.
-        assert!(
-            entry.tidal_config.is_none() || pfix_id.is_some(),
-            "add_source: tidal_config requires a planet-fixed frame \
-             (set rotation_model or t_inertial_pfix on the source)."
-        );
-
-        self.source_frame_ids.push(SourceFrameIds {
-            inertial: inertial_id,
-            pfix: pfix_id,
-        });
-        self.gravity_data.push(GravityData {
-            source: entry.source,
-            velocity: entry.velocity,
-            delta_c20: entry.delta_c20,
-            tidal_config: entry.tidal_config,
-            rotation_model: entry.rotation_model,
-            planet_omega: entry.planet_omega,
-        });
-        self.source_ephem_bodies.push(None);
-        idx
-    }
-
-    /// Configure ephemeris-based position updates for a source.
-    /// Each step, the source's position and velocity will be updated from DE4xx.
-    ///
-    /// `target` is the body this source represents (e.g., `EphemerisBody::Sun`).
-    /// `observer` is the integration frame center (e.g., `EphemerisBody::Earth`).
-    pub fn set_source_ephemeris(
-        &mut self,
-        source_idx: usize,
-        target: jeod_sim::EphemerisBody,
-        observer: jeod_sim::EphemerisBody,
-    ) {
-        assert!(
-            source_idx < self.source_ephem_bodies.len(),
-            "set_source_ephemeris: source_idx {source_idx} out of bounds (len = {})",
-            self.source_ephem_bodies.len()
-        );
-        // Root-frame conflict is caught by validate() → EphemerisOnRootSource.
-        // We don't panic here so that all misconfiguration errors are reported
-        // together in a single validate() pass rather than aborting on the first.
-        self.source_ephem_bodies[source_idx] = Some((target, observer));
-    }
-
-    /// Add a dynamic body from a [`VehicleConfig`]. Returns its index.
-    ///
-    /// The config is consumed and converted into internal state. Creates a
-    /// body frame in the frame tree under the integration frame. Use
-    /// [`body`](Simulation::body) to access results after stepping.
-    pub fn add_body(&mut self, config: VehicleConfig) -> usize {
-        let idx = self.bodies.len();
-
-        // Resolve integration frame from source index.
-        let integ_frame_id = config
-            .integ_source
-            .map(|src| {
-                self.source_frame_ids
-                    .get(src)
-                    .unwrap_or_else(|| {
-                        panic!(
-                            "VehicleConfig::integ_source index {src} is out of range; \
-                             {} source frame(s) configured",
-                            self.source_frame_ids.len()
-                        )
-                    })
-                    .inertial
-            })
-            .unwrap_or(self.root_frame_id);
-
-        // Create body frame in tree under the integration frame.
-        let body_frame_id = self.frame_tree.add_child(
-            integ_frame_id,
-            format!("body_{idx}.integ"),
-            RefFrameKind::Body,
-            RefFrameState {
-                trans: RefFrameTrans {
-                    position: config.trans.position,
-                    velocity: config.trans.velocity,
-                },
-                rot: RefFrameRot::default(),
-            },
-        );
-
-        self.bodies
-            .push(SimBody::from_config(config, integ_frame_id, body_frame_id));
-        idx
-    }
-
-    /// Read-only access to the reference frame tree.
-    pub fn frame_tree(&self) -> &FrameTree {
-        &self.frame_tree
-    }
-
-    /// Number of gravity sources.
-    pub fn num_sources(&self) -> usize {
-        self.gravity_data.len()
-    }
-
-    /// Get the inertial frame ID for a gravity source.
-    pub fn source_frame(&self, source_idx: usize) -> FrameId {
-        self.source_frame_ids
-            .get(source_idx)
-            .unwrap_or_else(|| {
-                panic!(
-                    "source_frame: source index {source_idx} is out of range; \
-                     {} source frame(s) configured",
-                    self.num_sources()
-                )
-            })
-            .inertial
-    }
-
-    /// Get the current position of a gravity source relative to the root
-    /// inertial frame. Returns `DVec3::ZERO` for the root-mapped central source.
-    pub fn source_position(&self, source_idx: usize) -> DVec3 {
-        let fid = self.source_frame(source_idx);
-        if fid == self.root_frame_id {
-            DVec3::ZERO
-        } else {
-            self.frame_tree.get(fid).state.trans.position
-        }
-    }
-
-    /// Set the position of a gravity source relative to the root inertial frame.
-    pub fn set_source_position(&mut self, source_idx: usize, position: DVec3) {
-        assert!(
-            source_idx < self.source_frame_ids.len(),
-            "set_source_position: source index {source_idx} out of range; \
-             {} source(s) configured",
-            self.source_frame_ids.len()
-        );
-        let fid = self.source_frame_ids[source_idx].inertial;
-        assert_ne!(
-            fid, self.root_frame_id,
-            "set_source_position: cannot set position of the root (central body) source"
-        );
-        self.frame_tree.get_mut(fid).state.trans.position = position;
-    }
-
-    /// Set the position and velocity of a gravity source relative to the root inertial frame.
-    ///
-    /// Prefer this over [`set_source_position`](Simulation::set_source_position)
-    /// when velocity is also available, to keep position and velocity consistent.
-    pub fn set_source_state(&mut self, source_idx: usize, position: DVec3, velocity: DVec3) {
-        assert!(
-            source_idx < self.source_frame_ids.len(),
-            "set_source_state: source index {source_idx} out of range; \
-             {} source(s) configured",
-            self.source_frame_ids.len()
-        );
-        let fid = self.source_frame_ids[source_idx].inertial;
-        assert_ne!(
-            fid, self.root_frame_id,
-            "set_source_state: cannot set state of the root (central body) source"
-        );
-        let node = self.frame_tree.get_mut(fid);
-        node.state.trans.position = position;
-        node.state.trans.velocity = velocity;
-        // Keep gravity_data velocity in sync for relativistic corrections (PPN).
-        self.gravity_data[source_idx].velocity = velocity;
-    }
-
-    /// Get the planet-fixed rotation matrix for a gravity source. Returns `None`
-    /// if the source has no rotation model (no pfix frame).
-    pub fn source_pfix_rotation(&self, source_idx: usize) -> Option<DMat3> {
-        self.source_frame_ids
-            .get(source_idx)
-            .unwrap_or_else(|| {
-                panic!(
-                    "source_pfix_rotation: source index {source_idx} out of range; \
-                     {} source(s) configured",
-                    self.source_frame_ids.len()
-                )
-            })
-            .pfix
-            .map(|pfix_id| self.frame_tree.get(pfix_id).state.rot.t_parent_this)
-    }
-
-    /// Get mutable access to a source's tidal configuration.
-    pub fn source_tidal_config_mut(
-        &mut self,
-        source_idx: usize,
-    ) -> Option<&mut jeod_gravity::tides::TidalConfig> {
-        let len = self.gravity_data.len();
-        self.gravity_data
-            .get_mut(source_idx)
-            .unwrap_or_else(|| {
-                panic!(
-                    "source_tidal_config_mut: source index {source_idx} out of range; \
-                     {len} source(s) configured",
-                )
-            })
-            .tidal_config
-            .as_mut()
-    }
-
-    /// Get the current ΔC20 tidal correction for a gravity source.
-    pub fn source_delta_c20(&self, source_idx: usize) -> f64 {
-        assert!(
-            source_idx < self.gravity_data.len(),
-            "source_delta_c20: source index {source_idx} out of range; \
-             {} source(s) configured",
-            self.gravity_data.len()
-        );
-        self.gravity_data[source_idx].delta_c20
-    }
-
-    /// Validate all bodies against JEOD invariants and apply auto-corrections.
-    ///
-    /// Call once before the first `step()`. Returns `Ok(())` if all bodies are
-    /// valid, or `Err(errors)` with all validation errors found.
-    ///
-    /// Also runs `GravityControl::check_validity()` on each control to
-    /// auto-correct degree/order (matching JEOD's `initialize_gravity_controls()`
-    /// and the Bevy adapter's startup validation).
-    // JEOD_INV: GV.03 — check_validity() called at startup (auto-corrections applied in-place)
-    pub fn validate(&mut self) -> Result<(), Vec<ValidationError>> {
-        let num_sources = self.gravity_data.len();
-        let mut all_errors = Vec::new();
-        for (body_idx, body) in self.bodies.iter_mut().enumerate() {
-            let plate_counts = body.flat_plate_state.as_ref().map(|fps| {
-                (
-                    fps.plates.len(),
-                    fps.temperatures.len(),
-                    fps.t_pow4_cached.len(),
-                )
-            });
-            let errors = jeod_sim::validate_body(
-                &body.config,
-                &body.gravity_controls,
-                true, // SimBody always has gravity_accel field
-                body.mass.as_ref(),
-                body.rot.is_some(),
-                Some(&body.trans),
-                |source_id: usize| self.gravity_data.get(source_id).map(|g| &g.source),
-                plate_counts,
-            );
-            all_errors.extend(errors);
-
-            // Validate shadow_body index
-            if let Some((idx, _radius)) = body.shadow_body {
-                if idx >= num_sources {
-                    all_errors.push(ValidationError::ShadowBodyOutOfRange {
-                        index: idx,
-                        num_sources,
-                    });
-                }
-            }
-
-            // Validate geodetic_planet index
-            if let Some((idx, _, _)) = body.geodetic_planet {
-                if idx >= num_sources {
-                    all_errors.push(ValidationError::GeodeticPlanetOutOfRange {
-                        index: idx,
-                        num_sources,
-                    });
-                }
-            }
-
-            // Validate orbital_elements_source index
-            if let Some(idx) = body.orbital_elements_source {
-                if idx >= num_sources {
-                    all_errors.push(ValidationError::OrbitalElementsSourceOutOfRange {
-                        index: idx,
-                        num_sources,
-                    });
-                }
-            }
-
-            // Validate force producers have mass (JEOD_INV: MA.01 — MassBody always present)
-            if (body.drag.is_some() || body.flat_plate_state.is_some()) && body.mass.is_none() {
-                all_errors.push(ValidationError::ForceProducerWithoutMass { body_idx });
-            }
-
-            // GaussJackson is translational-only (6-DOF not yet supported)
-            if matches!(
-                body.integrator,
-                jeod_dynamics::IntegratorType::GaussJackson(..)
-            ) && body.config.rotational_dynamics
-            {
-                all_errors.push(ValidationError::GaussJacksonWith6Dof { body_idx });
-            }
-
-            // ABM4 is translational-only (6-DOF not yet supported)
-            if matches!(body.integrator, jeod_dynamics::IntegratorType::Abm4)
-                && body.config.rotational_dynamics
-            {
-                all_errors.push(ValidationError::Abm4With6Dof { body_idx });
-            }
-
-            // GaussJackson config validation — delegates to GaussJacksonConfig::check()
-            // so the predicate is defined in one place.
-            if let jeod_dynamics::IntegratorType::GaussJackson(ref config) = body.integrator {
-                for detail in config.check() {
-                    all_errors
-                        .push(ValidationError::GaussJacksonConfigInvalid { body_idx, detail });
-                }
-            }
-
-            // Atmospheric state requires atmosphere config on the simulation
-            if body.atmospheric_state.is_some() && self.atmosphere.is_none() {
-                all_errors.push(ValidationError::AtmosphericStateWithoutAtmosphere { body_idx });
-            }
-
-            // Solar beta requires sun_source on the simulation
-            if body.compute_solar_beta && self.sun_source.is_none() {
-                all_errors.push(ValidationError::SolarBetaWithoutSunSource { body_idx });
-            }
-
-            // Earth lighting requires both sun_source and moon_source
-            if body.earth_lighting_config.is_some() {
-                if self.sun_source.is_none() {
-                    all_errors.push(ValidationError::EarthLightingWithoutSunSource { body_idx });
-                }
-                if self.moon_source.is_none() {
-                    all_errors.push(ValidationError::EarthLightingWithoutMoonSource { body_idx });
-                }
-            }
-
-            // Gravity torque requires both mass and rotational state
-            if body.compute_gravity_torque && (body.mass.is_none() || body.rot.is_none()) {
-                all_errors.push(ValidationError::GravityTorqueWithoutMassOrRot { body_idx });
-            }
-
-            // Frame switch target_source must be a valid source index AND
-            // present in the body's gravity controls (so the post-switch
-            // differential flip actually takes effect).
-            // Only validate active switches — JEOD only evaluates active switches.
-            for sw in &body.frame_switches {
-                if sw.active {
-                    let target = sw.target_source;
-                    if target >= num_sources {
-                        all_errors.push(ValidationError::FrameSwitchTargetSourceOutOfRange {
-                            body_idx,
-                            target_source: target,
-                            num_sources,
-                        });
-                    } else if !body
-                        .gravity_controls
-                        .controls
-                        .iter()
-                        .any(|c| c.source_name == target)
-                    {
-                        all_errors.push(ValidationError::FrameSwitchTargetSourceNotInControls {
-                            body_idx,
-                            target_source: target,
-                        });
-                    }
-                }
-            }
-
-            // Warn when body uses a non-root integration frame with features
-            // that assume root-inertial coordinates. JEOD evaluates these
-            // derived states in the central-body inertial frame; they will
-            // produce incorrect results in other frames.
-            {
-                let non_root_integ = body.integ_frame_id != self.root_frame_id;
-                let non_root_switch = body.frame_switches.iter().any(|sw| {
-                    sw.active
-                        && self
-                            .source_frame_ids
-                            .get(sw.target_source)
-                            .is_some_and(|frame| frame.inertial != self.root_frame_id)
-                });
-                if non_root_integ || non_root_switch {
-                    let has_root_dependent_feature = body.drag.is_some()
-                        || body.flat_plate_state.is_some()
-                        || body.cannonball_srp.is_some()
-                        || body.orbital_elements_source.is_some()
-                        || body.euler_sequence.is_some()
-                        || body.compute_lvlh
-                        || body.geodetic_planet.is_some()
-                        || body.compute_solar_beta
-                        || body.earth_lighting_config.is_some();
-                    if has_root_dependent_feature {
-                        all_errors.push(ValidationError::NonRootFrameWithRootDependentFeatures {
-                            body_idx,
-                        });
-                    }
-                }
-            }
-
-            // Apply gravity control auto-corrections (degree/order clamping).
-            // JEOD_INV: GV.03 — check_validity() auto-corrects out-of-range settings
-            for ctrl in &mut body.gravity_controls.controls {
-                if let Some(grav) = self.gravity_data.get(ctrl.source_name) {
-                    ctrl.check_validity(&grav.source);
-                }
-            }
-        }
-
-        // Validate sun_source index (simulation-level, outside body loop)
-        if let Some(idx) = self.sun_source {
-            if idx >= num_sources {
-                all_errors.push(ValidationError::SunSourceOutOfRange {
-                    index: idx,
-                    num_sources,
-                });
-            }
-        }
-
-        // Validate moon_source index
-        if let Some(idx) = self.moon_source {
-            if idx >= num_sources {
-                all_errors.push(ValidationError::MoonSourceOutOfRange {
-                    index: idx,
-                    num_sources,
-                });
-            }
-        }
-
-        // Validate atmosphere_planet_source index
-        if let Some(idx) = self.atmosphere_planet_source {
-            if idx >= num_sources {
-                all_errors.push(ValidationError::AtmospherePlanetOutOfRange {
-                    index: idx,
-                    num_sources,
-                });
-            }
-        }
-
-        // Ephemeris mapping on root-frame sources — would silently discard position.
-        for (i, ephem) in self.source_ephem_bodies.iter().enumerate() {
-            if ephem.is_some() && self.source_frame_ids[i].inertial == self.root_frame_id {
-                all_errors.push(ValidationError::EphemerisOnRootSource { source_idx: i });
-            }
-        }
-
-        // Contact pairs require the multi-body coupled RK4 path, which drives
-        // every body through the same kernel. Enforce RK4 + 6-DOF on all
-        // bodies (not just those that appear in a pair) so the field doc's
-        // "validation error" promise matches reality.
-        if !self.contact_pairs.is_empty() {
-            for (body_idx, body) in self.bodies.iter().enumerate() {
-                if !matches!(body.integrator, jeod_dynamics::IntegratorType::Rk4) {
-                    all_errors.push(ValidationError::ContactPairsRequireRk4 { body_idx });
-                }
-                if body.rot.is_none() || body.mass.is_none() {
-                    all_errors.push(ValidationError::ContactPairsRequire6Dof { body_idx });
-                }
-            }
-
-            // The coupled RK4 contact path evaluates `stage_trans`/`stage_rot`
-            // directly, and those stage states are in each body's integration
-            // frame. For pair-level consistency (and to match the no-transform
-            // convention in `evaluate_contact_pair`), both bodies must share
-            // the same integration frame, and that frame must be the root
-            // inertial frame.
-            for (pair_idx, pair) in self.contact_pairs.iter().enumerate() {
-                let frame_a = self.bodies[pair.body_a].integ_frame_id;
-                let frame_b = self.bodies[pair.body_b].integ_frame_id;
-                if frame_a != frame_b {
-                    all_errors.push(ValidationError::ContactPairFrameMismatch {
-                        pair_idx,
-                        body_a: pair.body_a,
-                        body_b: pair.body_b,
-                        frame_a,
-                        frame_b,
-                    });
-                    continue;
-                }
-                if frame_a != self.root_frame_id {
-                    // frame_a == frame_b at this point, so both bodies
-                    // share the same non-root frame. Emit one error per
-                    // body so debuggers see the full set.
-                    all_errors.push(ValidationError::ContactPairNonRootFrame {
-                        pair_idx,
-                        body_idx: pair.body_a,
-                        frame: frame_a,
-                        root: self.root_frame_id,
-                    });
-                    all_errors.push(ValidationError::ContactPairNonRootFrame {
-                        pair_idx,
-                        body_idx: pair.body_b,
-                        frame: frame_a,
-                        root: self.root_frame_id,
-                    });
-                }
-            }
-        }
-
-        // Separate warnings from fatal errors — warnings are logged, not returned.
-        let mut fatal = Vec::new();
-        for error in all_errors {
-            if error.is_warning() {
-                log::warn!("{error}");
-            } else {
-                fatal.push(error);
-            }
-        }
-        if !fatal.is_empty() {
-            return Err(fatal);
-        }
-
-        // Auto-initialize Gauss-Jackson state for bodies that need it.
-        // Check config consistency for pre-existing states.
-        for (body_idx, body) in self.bodies.iter_mut().enumerate() {
-            if let jeod_dynamics::IntegratorType::GaussJackson(ref config) = body.integrator {
-                match &body.gj_state {
-                    None => {
-                        body.gj_state = Some(jeod_dynamics::GaussJacksonState::new(*config));
-                    }
-                    Some(state) if state.config() != config => {
-                        fatal.push(ValidationError::GaussJacksonConfigInvalid {
-                            body_idx,
-                            detail: format!(
-                                "existing gj_state config does not match IntegratorType config \
-                                 (initial_order {}/{}, final_order {}/{}). \
-                                 Remove gj_state or recreate from the same config.",
-                                state.config().initial_order,
-                                config.initial_order,
-                                state.config().final_order,
-                                config.final_order,
-                            ),
-                        });
-                    }
-                    Some(_) => {} // config matches, keep existing state
-                }
-            }
-            // Auto-initialize ABM4 state for bodies that need it.
-            if matches!(body.integrator, jeod_dynamics::IntegratorType::Abm4)
-                && body.abm4_state.is_none()
-            {
-                body.abm4_state = Some(jeod_dynamics::Abm4State::new());
-            }
-        }
-        if !fatal.is_empty() {
-            return Err(fatal);
-        }
-
-        Ok(())
-    }
-
-    /// Advance the simulation by one timestep.
-    ///
-    /// Runs the full JEOD pipeline in order:
-    /// 1. Time update
-    /// 2. Ephemeris update (planet-fixed rotations + frame tree sync)
-    /// 3. Mass update (recompute derived quantities)
-    /// 4. Gravity computation
-    /// 5. Atmosphere evaluation
-    /// 6. Interaction computation (drag, SRP, gravity torque)
-    /// 7. Force collection and frame derivative computation
-    /// 8. State integration (RK4, with sub-stage tree updates)
-    /// 9. Derived state computation
-    pub fn step(&mut self) -> Result<(), StepError> {
-        self.step_internal(self.dt)
-    }
-
-    /// Get the position and velocity of a frame relative to the root inertial frame.
-    pub fn frame_origin(&self, frame_id: FrameId) -> (DVec3, DVec3) {
-        if frame_id == self.root_frame_id {
-            return (DVec3::ZERO, DVec3::ZERO);
-        }
-        let state = self
-            .frame_tree
-            .compute_relative_state(self.root_frame_id, frame_id);
-        (state.trans.position, state.trans.velocity)
-    }
-
-    /// Sync a planet-fixed frame node's rotation state from a computed matrix.
-    ///
-    /// Sets `t_parent_this`, derives `q_parent_this` from it, and sets
-    /// `ang_vel_this = [0, 0, planet_omega]` matching JEOD's `planet_rnp.cc`.
-    /// The `planet_omega` value comes from [`PlanetConfig::omega`] via
-    /// [`GravityData::planet_omega`].
-    fn sync_pfix_rotation(
-        frame_tree: &mut jeod_frames::FrameTree,
-        pfix_id: jeod_frames::FrameId,
-        rotation: DMat3,
-        planet_omega: f64,
-    ) {
-        let node = frame_tree.get_mut(pfix_id);
-        node.state.rot.t_parent_this = rotation;
-        node.state.rot.q_parent_this = JeodQuat::left_quat_from_transformation(&rotation);
-        // JEOD sets ang_vel_this = [0, 0, planet_omega] in planet_rnp.cc.
-        // This is used by compute_relative_state velocity composition.
-        node.state.rot.ang_vel_this = DVec3::new(0.0, 0.0, planet_omega);
-    }
-
-    /// Internal step with explicit dt (avoids temporary mutation of `self.dt`
-    /// in `step_until`).
-    fn step_internal(&mut self, dt: f64) -> Result<(), StepError> {
-        // ── 1. Time update ──
-        self.time.advance(dt);
-
-        // ── 2. Ephemeris update — planet-fixed rotations + frame tree sync ──
-        // JEOD_INV: DM.13 — ephemeris updated before gravity
-        // Per-source rotation dispatch: each source has its own rotation model.
-        // Lazy-compute Earth RNP only if needed (most common case).
-        let mut earth_rotation: Option<DMat3> = Option::None;
-        for (i, grav) in self.gravity_data.iter_mut().enumerate() {
-            match grav.rotation_model {
-                RotationModel::None => {}
-                RotationModel::EarthRNP => {
-                    let rotation = *earth_rotation.get_or_insert_with(|| {
-                        jeod_sim::compute_t_parent_this_from_tjt_with_polar(
-                            self.time.gmst_seconds,
-                            self.time.tt_tjt(),
-                            self.polar_motion,
-                        )
-                    });
-                    // Sync to frame tree pfix node.
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-                RotationModel::MarsIAU => {
-                    // JEOD's RNPMars receives TT seconds since J2000 (time_tt.seconds).
-                    let tt_s_since_j2000 = (self.time.tt_tjt() - jeod_time::epoch::J2000_TT_TJT)
-                        * jeod_time::epoch::SECONDS_PER_DAY;
-                    let rotation =
-                        jeod_frames::rotation_mars::compute_mars_rotation(tt_s_since_j2000);
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-                RotationModel::MoonIAU => {
-                    let tdb_jd = self.time.tdb_julian_date();
-                    let tdb_s_since_j2000 = (tdb_jd - jeod_time::epoch::J2000_TT_JD)
-                        * jeod_time::epoch::SECONDS_PER_DAY;
-                    let rotation =
-                        jeod_frames::rotation_moon::compute_moon_rotation(tdb_s_since_j2000);
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-                RotationModel::MoonDE421 => {
-                    let eph = self.ephemeris.as_ref().expect(
-                        "MoonDE421 rotation requires ephemeris with BPC. \
-                         Set sim.ephemeris = Some(eph) after calling eph.load_bpc().",
-                    );
-                    let tdb_jd = self.time.tdb_julian_date();
-                    let rotation = eph
-                        .get_body_rotation(jeod_sim::EphemerisBody::Moon, tdb_jd)
-                        .expect("Moon DE421 BPC rotation query failed");
-                    if let Some(pfix_id) = self.source_frame_ids[i].pfix {
-                        Self::sync_pfix_rotation(
-                            &mut self.frame_tree,
-                            pfix_id,
-                            rotation,
-                            grav.planet_omega,
-                        );
-                    }
-                }
-            }
-            // Compute tidal ΔC20 if configured; otherwise clear any stale value.
-            if let Some(ref config) = grav.tidal_config {
-                let pfix_id = self.source_frame_ids[i]
-                    .pfix
-                    .expect("tidal_config requires a planet-fixed frame (set rotation_model or t_inertial_pfix).");
-                let rotation = self.frame_tree.get(pfix_id).state.rot.t_parent_this;
-                grav.delta_c20 = jeod_gravity::tides::compute_delta_c20(config, &rotation);
-            } else {
-                grav.delta_c20 = 0.0;
-            }
-        }
-
-        // ── 2b. Ephemeris update — source positions from DE4xx ──
-        // Update source positions from ephemeris each step and sync to frame tree.
-        if let Some(ref eph) = self.ephemeris {
-            let tdb_jd = self.time.tdb_julian_date();
-            for i in 0..self.source_ephem_bodies.len() {
-                if let Some(Some((target, observer))) = self.source_ephem_bodies.get(i) {
-                    let (pos_typed, vel_typed) = eph
-                        .get_state_typed(*target, *observer, tdb_jd)
-                        .map_err(|e| StepError::EphemerisLookup {
-                            source_idx: i,
-                            target: *target,
-                            observer: *observer,
-                            tdb_jd,
-                            message: e.to_string(),
-                        })?;
-                    let (pos, vel) = (pos_typed.raw_si(), vel_typed.raw_si());
-                    // Root-mapped sources cannot consume ephemeris position updates:
-                    // the root frame must remain identity, so accepting such a
-                    // mapping would silently ignore `pos` and yield an incorrect
-                    // source position.
-                    let fid = self.source_frame_ids[i].inertial;
-                    assert!(
-                        fid != self.root_frame_id,
-                        "Invalid ephemeris mapping for source {i} \
-                         ({target:?} wrt {observer:?}): source inertial frame is the root frame, \
-                         whose state must remain identity. Root-mapped sources cannot use \
-                         ephemeris position updates."
-                    );
-                    // Update frame tree node with ephemeris position/velocity.
-                    let node = self.frame_tree.get_mut(fid);
-                    node.state.trans.position = pos;
-                    node.state.trans.velocity = vel;
-                    // Also update gravity_data velocity for relativistic corrections.
-                    self.gravity_data[i].velocity = vel;
-                }
-            }
-        }
-
-        // ── 3. Mass update — recompute inverse_mass/inverse_inertia ──
-        for body in &mut self.bodies {
-            if let Some(ref mut mass) = body.mass {
-                mass.recompute_derived();
-            }
-        }
-
-        // Precompute frame origins from the tree for all body integration frames.
-        let body_integ_origins: Vec<(DVec3, DVec3)> = self
-            .bodies
-            .iter()
-            .map(|b| self.frame_origin(b.integ_frame_id))
-            .collect();
-
-        // ── 4. Environment — gravity ──
-        // Helper: resolve source to gravity data via frame tree.
-        let gravity_data = &self.gravity_data;
-        let source_frame_ids = &self.source_frame_ids;
-        let frame_tree = &self.frame_tree;
-        let root_fid = self.root_frame_id;
-        let resolve_source = |source_id: usize| -> Option<jeod_sim::ResolvedSource<'_>> {
-            let grav = gravity_data.get(source_id)?;
-            let sfids = &source_frame_ids[source_id];
-            let src_node = frame_tree.get(sfids.inertial);
-            let position = if sfids.inertial == root_fid {
-                DVec3::ZERO
-            } else {
-                src_node.state.trans.position
-            };
-            let rotation = sfids
-                .pfix
-                .map(|pfix_id| &frame_tree.get(pfix_id).state.rot.t_parent_this);
-            Some(jeod_sim::ResolvedSource {
-                source: &grav.source,
-                rotation,
-                position,
-                delta_c20: grav.delta_c20,
-                has_delta_coeffs: grav.tidal_config.is_some(),
-            })
-        };
-
-        for (body_idx, body) in self.bodies.iter_mut().enumerate() {
-            let integ_origin = body_integ_origins[body_idx].0;
-            body.gravity_accel = accumulate_gravity(
-                body.trans.position + integ_origin,
-                &body.gravity_controls,
-                integ_origin,
-                resolve_source,
-            );
-        }
-
-        // ── 4b. Relativistic corrections ──
-        // After Newtonian gravity, apply post-Newtonian PPN correction for
-        // any source with `relativistic: true`. Folkner eq 27 (β=γ=1).
-        // PPN uses inertial coordinates — convert from integration frame.
-        let resolve_rel_source =
-            |source_id: usize| -> Option<jeod_sim::ResolvedRelativisticSource> {
-                let grav = gravity_data.get(source_id)?;
-                let sfids = &source_frame_ids[source_id];
-                let position = if sfids.inertial == root_fid {
-                    DVec3::ZERO
-                } else {
-                    frame_tree.get(sfids.inertial).state.trans.position
-                };
-                Some(jeod_sim::ResolvedRelativisticSource {
-                    mu: grav.source.mu,
-                    position,
-                    // Use velocity from gravity_data, not the tree node, because
-                    // central bodies at the root frame have zero tree velocity
-                    // but may have physical velocity for relativistic corrections.
-                    velocity: grav.velocity,
-                })
-            };
-
-        for (body_idx, body) in self.bodies.iter_mut().enumerate() {
-            let (origin, origin_vel) = body_integ_origins[body_idx];
-            body.gravity_accel.grav_accel += jeod_sim::accumulate_relativistic_corrections(
-                body.trans.position + origin,
-                body.trans.velocity + origin_vel,
-                &body.gravity_controls,
-                resolve_rel_source,
-            );
-        }
-
-        // ── 5. Environment — atmosphere ──
-        if let Some(ref atmos_config) = self.atmosphere {
-            let t_pfix = self
-                .atmosphere_planet_source
-                .and_then(|idx| self.source_frame_ids.get(idx))
-                .and_then(|sfids| sfids.pfix)
-                .map(|pfix_id| &self.frame_tree.get(pfix_id).state.rot.t_parent_this);
-            let tai_tjt = Some(self.time.tai_tjt);
-
-            for body in &mut self.bodies {
-                if body.atmospheric_state.is_some() {
-                    body.atmospheric_state = Some(evaluate_atmosphere(
-                        atmos_config,
-                        body.trans.position,
-                        t_pfix,
-                        tai_tjt,
-                    ));
-                }
-            }
-        }
-
-        // ── 6. Interactions — drag, SRP, gravity torque ──
-        // sun_pos is also used in stage 9 (solar beta, earth lighting); compute once here.
-        let sun_pos = self.sun_source.map(|idx| self.source_position(idx));
-        let moon_pos = self.moon_source.map(|idx| self.source_position(idx));
-        let source_frame_ids = &self.source_frame_ids;
-        let frame_tree = &self.frame_tree;
-        let root_fid = self.root_frame_id;
-
-        for body in &mut self.bodies {
-            // Compute structural transform once (shared by drag and flat-plate SRP)
-            let t_inertial_body = body.rot.as_ref().map_or(DMat3::IDENTITY, |r| {
-                r.quaternion.left_quat_to_transformation()
-            });
-            let t_inertial_struct =
-                jeod_sim::compute_t_inertial_struct(&body.t_struct_body, &t_inertial_body);
-
-            // Aerodynamic drag
-            body.aero_force = None;
-            if let (Some(ref drag_config), Some(ref atmos)) = (&body.drag, &body.atmospheric_state)
-            {
-                body.aero_force = Some(jeod_sim::compute_drag(
-                    drag_config,
-                    atmos,
-                    body.trans.velocity,
-                    body.rot.as_ref(),
-                    body.t_struct_body,
-                ));
-            }
-
-            // Solar radiation pressure (flat-plate)
-            body.radiation_force = None;
-            if let Some(ref mut fps) = body.flat_plate_state {
-                fps.stage_inputs = None;
-            }
-            if let Some(sun_position) = sun_pos {
-                if let Some(ref mut fps) = body.flat_plate_state {
-                    // Flat-plate SRP with thermal emission
-                    let sun_to_vehicle = body.trans.position - sun_position;
-                    let distance = sun_to_vehicle.length();
-                    // Skip SRP (not the whole body) if too close to Sun
-                    if distance >= 1.0 {
-                        let flux_inertial_hat = sun_to_vehicle / distance;
-                        let flux_mag = jeod_sim::solar_flux_at_distance(distance);
-
-                        // Shadow fraction (step-constant — matches JEOD's
-                        // scheduled-class shadow evaluation in SIM_3_ORBIT).
-                        let illum_factor = body
-                            .shadow_body
-                            .map(|(idx, radius)| {
-                                jeod_sim::compute_shadow_fraction(
-                                    body.trans.position,
-                                    sun_position,
-                                    {
-                                        let fid = source_frame_ids[idx].inertial;
-                                        if fid == root_fid {
-                                            DVec3::ZERO
-                                        } else {
-                                            frame_tree.get(fid).state.trans.position
-                                        }
-                                    },
-                                    radius,
-                                    jeod_sim::SOLAR_RADIUS,
-                                )
-                            })
-                            .unwrap_or(1.0);
-
-                        let center_grav = body.mass.as_ref().map_or(DVec3::ZERO, |m| m.position);
-
-                        match fps.integration_order {
-                            jeod_sim::ThermalIntegrationOrder::Scheduled => {
-                                // Scheduled-class: compute SRP force + Euler T
-                                // update once per step (JEOD SIM_3_ORBIT).
-                                let flux_struct_hat = t_inertial_struct * flux_inertial_hat;
-                                let srp_result = jeod_sim::compute_flat_plate_srp_thermal(
-                                    &fps.plates,
-                                    &fps.t_pow4_cached,
-                                    flux_struct_hat,
-                                    flux_mag,
-                                    center_grav,
-                                    illum_factor,
-                                );
-
-                                // Force: structural → inertial. Torque: stays structural.
-                                let force_inertial =
-                                    t_inertial_struct.transpose() * srp_result.force;
-                                body.radiation_force = Some(RadiationForce {
-                                    force: force_inertial,
-                                    torque: srp_result.torque,
-                                });
-
-                                fps.integrate_temperatures(&srp_result.temp_dots, dt);
-                            }
-                            jeod_sim::ThermalIntegrationOrder::DerivativeFirstOrder
-                            | jeod_sim::ThermalIntegrationOrder::DerivativeRk4 => {
-                                // Derivative-class: SRP force (and optionally T)
-                                // recomputed per RK4 stage. Stash the step-start
-                                // inputs on the plate state; Stage 8 consumes
-                                // them via `integrate_body_coupled` below.
-                                fps.stage_inputs = Some(jeod_sim::FlatPlateStageInputs {
-                                    sun_position,
-                                    illum_factor,
-                                    center_grav,
-                                });
-                                // `radiation_force` is left None here; Stage 8
-                                // writes a representative final-stage value so
-                                // `VehicleOutput` still reports SRP force.
-                            }
-                        }
-                    }
-                } else if let Some((cx_area, albedo, diffuse)) = body.cannonball_srp {
-                    let illum_factor = body
-                        .shadow_body
-                        .map(|(idx, radius)| {
-                            jeod_sim::compute_shadow_fraction(
-                                body.trans.position,
-                                sun_position,
-                                {
-                                    let fid = source_frame_ids[idx].inertial;
-                                    if fid == root_fid {
-                                        DVec3::ZERO
-                                    } else {
-                                        frame_tree.get(fid).state.trans.position
-                                    }
-                                },
-                                radius,
-                                jeod_sim::SOLAR_RADIUS,
-                            )
-                        })
-                        .unwrap_or(1.0);
-
-                    let force = jeod_sim::compute_cannonball_srp(
-                        body.trans.position,
-                        sun_position,
-                        cx_area,
-                        albedo,
-                        diffuse,
-                        illum_factor,
-                    );
-                    if force != DVec3::ZERO {
-                        body.radiation_force = Some(RadiationForce {
-                            force,
-                            torque: DVec3::ZERO,
-                        });
-                    }
-                }
-            }
-
-            // Gravity gradient torque
-            body.gravity_torque = None;
-            if body.compute_gravity_torque {
-                if let (Some(ref rot), Some(ref mass)) = (&body.rot, &body.mass) {
-                    body.gravity_torque = Some(jeod_sim::compute_gravity_torque(
-                        &body.gravity_accel.grav_grad,
-                        rot,
-                        &mass.inertia,
-                    ));
-                }
-            }
-        }
-
-        // ── 7. Force collection ──
-        for body in &mut self.bodies {
-            let (total, derivs) = collect_and_resolve_forces(
-                body.aero_force.as_ref(),
-                body.radiation_force.as_ref(),
-                body.gravity_torque,
-                body.rot.as_ref(),
-                body.t_struct_body,
-                body.mass.as_ref(),
-                body.gravity_accel.grav_accel,
-            );
-            body.total_force = total;
-            body.frame_derivs = derivs;
-
-            // Apply external force/torque (set by caller between steps).
-            // Recompute frame derivatives so they stay consistent with total_force.
-            body.total_force.force += body.external_force;
-            body.total_force.torque += body.external_torque;
-            if body.external_force != DVec3::ZERO {
-                if let Some(mass) = &body.mass {
-                    body.frame_derivs.trans_accel += body.external_force * mass.inverse_mass;
-                }
-            }
-            if body.external_torque != DVec3::ZERO {
-                if let Some(mass) = &body.mass {
-                    body.frame_derivs.rot_accel += mass.inverse_inertia * body.external_torque;
-                }
-            }
-        }
-
-        // ── 8. Integration ──
-        // Gravity (including relativistic corrections) is recomputed at each
-        // RK4 intermediate state for 4th-order accuracy, matching JEOD's
-        // DynamicsIntegrationGroup where the derivative function calls gravity
-        // at every stage with the current intermediate position and velocity.
-        //
-        // For RK4 sub-stage evaluations, source positions are derived from a
-        // linear interpolation of their base inertial position using
-        // velocity * (time_frac * dt), matching JEOD's behavior of evaluating
-        // gravity using the current sub-stage source state.
-        //
-        // Snapshot base source positions and velocities for sub-stage interpolation.
-        let base_positions: Vec<DVec3> = self
-            .source_frame_ids
-            .iter()
-            .map(|sfids| {
-                if sfids.inertial == self.root_frame_id {
-                    DVec3::ZERO
-                } else {
-                    self.frame_tree.get(sfids.inertial).state.trans.position
-                }
-            })
-            .collect();
-        let base_velocities: Vec<DVec3> = self
-            .source_frame_ids
-            .iter()
-            .map(|sfids| {
-                if sfids.inertial == self.root_frame_id {
-                    DVec3::ZERO
-                } else {
-                    self.frame_tree.get(sfids.inertial).state.trans.velocity
-                }
-            })
-            .collect();
-
-        let gravity_data = &self.gravity_data;
-        let source_frame_ids = &self.source_frame_ids;
-        let frame_tree = &self.frame_tree;
-        let root_fid = self.root_frame_id;
-
-        // Precompute per-body relativistic "other source" lists outside the
-        // closures to avoid heap allocation at every RK4 stage.
-        // Indexed by body_idx.
-        type RelDatum = (
-            f64,
-            DVec3,
-            DVec3,
-            Vec<jeod_gravity::relativistic::RelativisticSource>,
-        );
-        let per_body_rel_data: Vec<Vec<RelDatum>> = self
-            .bodies
-            .iter()
-            .map(|body| {
-                let controls = &body.gravity_controls;
-                controls
-                    .controls
-                    .iter()
-                    .filter(|c| c.relativistic)
-                    .filter_map(|ctrl| {
-                        let grav = gravity_data.get(ctrl.source_name)?;
-                        let sfids = &source_frame_ids[ctrl.source_name];
-                        let src_pos = if sfids.inertial == root_fid {
-                            DVec3::ZERO
-                        } else {
-                            frame_tree.get(sfids.inertial).state.trans.position
-                        };
-                        let src_vel = grav.velocity;
-                        let other: Vec<_> = controls
-                            .controls
-                            .iter()
-                            .filter(|c| c.source_name != ctrl.source_name)
-                            .filter_map(|c| {
-                                let g = gravity_data.get(c.source_name)?;
-                                let sf = &source_frame_ids[c.source_name];
-                                let pos = if sf.inertial == root_fid {
-                                    DVec3::ZERO
-                                } else {
-                                    frame_tree.get(sf.inertial).state.trans.position
-                                };
-                                Some(jeod_gravity::relativistic::RelativisticSource {
-                                    mu: g.source.mu,
-                                    position: pos,
-                                })
-                            })
-                            .collect();
-                        Some((grav.source.mu, src_pos, src_vel, other))
-                    })
-                    .collect()
-            })
-            .collect();
-
-        // Dynamic timestep: JEOD's `integ_dyndt = sim_dt * time_scale_factor`
-        // (`standard_integration_controls.cc:80-82`). Integrators step in
-        // dynamic time, so stage-time interpolations of the integration
-        // frame origin and source positions must scale by `integ_dt`, not
-        // by the raw `dt` — otherwise reversed/scaled time produces
-        // inconsistent gravity during coupled integration.
-        let integ_dt = dt * self.time.time_scale_factor;
-
-        // Helper: evaluate gravity (Newtonian + relativistic) at an
-        // intermediate (pos, vel) for the given body. Takes `controls` as
-        // a parameter so the non-contact path can borrow
-        // `&body.gravity_controls` directly (no per-step clone), while
-        // the contact-coupled path can pass slices of its cloned
-        // `per_body_gravity_controls` snapshot.
-        let eval_body_gravity = |controls: &GravityControls<usize>,
-                                 body_idx: usize,
-                                 pos: DVec3,
-                                 vel: DVec3,
-                                 time_frac: f64|
-         -> DVec3 {
-            let (integ_origin, integ_vel) = body_integ_origins[body_idx];
-            let stage_dt = time_frac * integ_dt;
-            let origin = integ_origin + integ_vel * stage_dt;
-            let sub_dt = if integ_vel != DVec3::ZERO {
-                stage_dt
-            } else {
-                0.0
-            };
-            let mut accel =
-                accumulate_gravity(pos + origin, controls, origin, |source_id: usize| {
-                    let grav = gravity_data.get(source_id)?;
-                    let sfids = &source_frame_ids[source_id];
-                    let position = base_positions[source_id] + base_velocities[source_id] * sub_dt;
-                    let rotation = sfids
-                        .pfix
-                        .map(|pfix_id| &frame_tree.get(pfix_id).state.rot.t_parent_this);
-                    Some(jeod_sim::ResolvedSource {
-                        source: &grav.source,
-                        rotation,
-                        position,
-                        delta_c20: grav.delta_c20,
-                        has_delta_coeffs: grav.tidal_config.is_some(),
-                    })
-                })
-                .grav_accel;
-            let pos_eci = pos + origin;
-            let vel_eci = vel + integ_vel;
-            for &(mu, src_pos, src_vel, ref other) in &per_body_rel_data[body_idx] {
-                accel += jeod_gravity::relativistic::compute_relativistic_correction(
-                    mu, src_pos, pos_eci, vel_eci, src_vel, other,
-                );
-            }
-            accel
-        };
-
-        if self.contact_pairs.is_empty() {
-            // ── Standard path: per-body RK4 / GJ integration ──
-            // No clone of gravity_controls: the outer iter_mut gives us
-            // a &mut SimBody, and Rust's disjoint-field split borrow lets
-            // the closure capture &body.gravity_controls while other
-            // fields of `body` are borrowed mutably for the integrator.
-            let time_scale_factor = self.time.time_scale_factor;
-            for (body_idx, body) in self.bodies.iter_mut().enumerate() {
-                let stage_inputs_and_order = body
-                    .flat_plate_state
-                    .as_ref()
-                    .and_then(|fps| fps.stage_inputs.map(|si| (si, fps.integration_order)));
-                if let Some((srp_inputs, thermal_order)) = stage_inputs_and_order {
-                    // JEOD_INV: IN.32 — derivative-class thermal: SRP force
-                    // (and temp_dots for DerivativeRk4) recomputed at each
-                    // RK4 stage from the intermediate orbital + thermal
-                    // state. Aero / external / gravity-gradient torque are
-                    // step-constant (scheduled-class) and captured once
-                    // from total_force.
-                    assert!(
-                        matches!(body.integrator, jeod_dynamics::IntegratorType::Rk4),
-                        "ThermalIntegrationOrder::{thermal_order:?} requires \
-                         jeod_dynamics::IntegratorType::Rk4 for body {body_idx}; \
-                         switch the body integrator to Rk4, or choose \
-                         ThermalIntegrationOrder::Scheduled to avoid the coupled \
-                         RK4 thermal path.",
-                    );
-                    let t_struct_body = body.t_struct_body;
-                    let non_grav_non_srp_force = body.total_force.force;
-                    let constant_torque = body.total_force.torque;
-                    let config = body.config;
-                    let controls = &body.gravity_controls;
-                    // Stash the final-stage SRP result so we can write a
-                    // representative `radiation_force` for `VehicleOutput`.
-                    let mut final_srp_inertial_force = DVec3::ZERO;
-                    let mut final_srp_torque = DVec3::ZERO;
-                    // For DerivativeFirstOrder: capture stage-1 temp_dots and
-                    // feed them back at stages 2-4, collapsing the RK4 thermal
-                    // combine to Euler (matches JEOD's ER7_Utils first-order
-                    // integrator behavior while still evaluating SRP per stage
-                    // for the orbital RK4).
-                    let mut k1_temp_dots: Option<Vec<f64>> = None;
-                    let mass_copy = body.mass;
-
-                    integrate_body_coupled(
-                        &config,
-                        &mut body.trans,
-                        body.rot.as_mut(),
-                        mass_copy.as_ref(),
-                        |stage_trans, stage_rot, stage_thermal, time_frac| {
-                            let gravity_accel = eval_body_gravity(
-                                controls,
-                                body_idx,
-                                stage_trans.position,
-                                stage_trans.velocity,
-                                time_frac,
-                            );
-                            let t_inertial_body = stage_rot.map_or(DMat3::IDENTITY, |r| {
-                                r.quaternion.left_quat_to_transformation()
-                            });
-                            let t_inertial_struct = jeod_sim::compute_t_inertial_struct(
-                                &t_struct_body,
-                                &t_inertial_body,
-                            );
-                            // Per-stage flux recompute from intermediate vehicle
-                            // position — matches JEOD's derivative-class
-                            // `RadiationSource::calculate_flux`. Sun position is
-                            // step-constant (ephemeris is scheduled-class).
-                            let sun_to_vehicle = stage_trans.position - srp_inputs.sun_position;
-                            let distance = sun_to_vehicle.length().max(1.0);
-                            let stage_flux_inertial_hat = sun_to_vehicle / distance;
-                            let stage_flux_mag = jeod_sim::solar_flux_at_distance(distance);
-                            let flux_struct_hat = t_inertial_struct * stage_flux_inertial_hat;
-                            let srp_result = jeod_sim::compute_flat_plate_srp_thermal(
-                                &stage_thermal.plates,
-                                &stage_thermal.t_pow4_cached,
-                                flux_struct_hat,
-                                stage_flux_mag,
-                                srp_inputs.center_grav,
-                                srp_inputs.illum_factor,
-                            );
-                            let srp_force_inertial =
-                                t_inertial_struct.transpose() * srp_result.force;
-                            // Stage 4 (time_frac == 1.0) is the representative
-                            // final-state SRP — cached for writeback below.
-                            final_srp_inertial_force = srp_force_inertial;
-                            final_srp_torque = srp_result.torque;
-                            let temp_dots = match thermal_order {
-                                jeod_sim::ThermalIntegrationOrder::DerivativeRk4 => {
-                                    srp_result.temp_dots
-                                }
-                                jeod_sim::ThermalIntegrationOrder::DerivativeFirstOrder => {
-                                    // Capture at stage 1 (time_frac == 0.0);
-                                    // reuse at stages 2-4 so RK4 combine
-                                    // collapses to Euler over k1.
-                                    if time_frac == 0.0 {
-                                        k1_temp_dots = Some(srp_result.temp_dots.clone());
-                                        srp_result.temp_dots
-                                    } else {
-                                        k1_temp_dots
-                                            .as_ref()
-                                            .expect("stage 1 runs before stages 2-4")
-                                            .clone()
-                                    }
-                                }
-                                jeod_sim::ThermalIntegrationOrder::Scheduled => {
-                                    unreachable!(
-                                        "Scheduled thermal bodies do not enter the coupled path"
-                                    )
-                                }
-                            };
-                            // `srp_result.torque` is structural-frame per
-                            // `FlatPlateSrpResult` docs; `constant_torque`
-                            // is body-frame (from `collect_and_resolve_forces`).
-                            // Rotate to body frame before summing so the
-                            // coupled integrator's rotational dynamics
-                            // are correct when t_struct_body != IDENTITY.
-                            let srp_torque_body = t_struct_body * srp_result.torque;
-                            CoupledStageEval {
-                                gravity_accel,
-                                non_grav_force: non_grav_non_srp_force + srp_force_inertial,
-                                torque: constant_torque + srp_torque_body,
-                                temp_dots,
-                            }
-                        },
-                        body.flat_plate_state
-                            .as_mut()
-                            .expect("srp_stage_inputs implies flat_plate_state"),
-                        dt,
-                        time_scale_factor,
-                    );
-
-                    body.radiation_force = Some(RadiationForce {
-                        force: final_srp_inertial_force,
-                        torque: final_srp_torque,
-                    });
-                    // Backfill `TotalForce` and `FrameDerivatives` with the
-                    // final-stage SRP contribution so downstream observers
-                    // reading these see SRP-inclusive values — matching the
-                    // Scheduled-mode invariant that `total_force` reflects
-                    // every applied force and `frame_derivs` the resulting
-                    // accelerations. In derivative modes this is a
-                    // "representative stage" (stage 4) snapshot, same as
-                    // `radiation_force` above.
-                    body.total_force.force += final_srp_inertial_force;
-                    let final_srp_torque_body = t_struct_body * final_srp_torque;
-                    body.total_force.torque += final_srp_torque_body;
-                    if let Some(mass) = body.mass {
-                        body.frame_derivs.trans_accel +=
-                            final_srp_inertial_force * mass.inverse_mass;
-                        body.frame_derivs.rot_accel += mass.inverse_inertia * final_srp_torque_body;
-                    }
-                } else {
-                    let controls = &body.gravity_controls;
-                    integrate_body(
-                        &body.config,
-                        &mut body.trans,
-                        body.rot.as_mut(),
-                        body.mass.as_ref(),
-                        |pos, vel, time_frac| {
-                            eval_body_gravity(controls, body_idx, pos, vel, time_frac)
-                        },
-                        body.total_force.force,
-                        body.total_force.torque,
-                        dt,
-                        time_scale_factor,
-                        body.integrator,
-                        body.gj_state.as_mut(),
-                        body.abm4_state.as_mut(),
-                    );
-                }
-            }
-        } else {
-            // ── Contact-coupled path: multi-body RK4 where contact forces
-            //    are recomputed at each stage from all bodies' intermediate
-            //    states (matching JEOD's `check_contact()` derivative job).
-            // JEOD_INV: IN.31 — contact evaluated at every derivative evaluation
-            //
-            // Enforce preconditions: all bodies participating in contact
-            // pairs must use RK4 + 6-DOF. We integrate ALL bodies through
-            // the coupled path; bodies without contact pairs just pass
-            // their constant forces through the same RK4 kernel.
-            assert!(
-                self.bodies
-                    .iter()
-                    .all(|b| matches!(b.integrator, jeod_dynamics::IntegratorType::Rk4)),
-                "contact pairs require RK4 integrator on all bodies"
-            );
-            assert!(
-                self.bodies
-                    .iter()
-                    .all(|b| b.rot.is_some() && b.mass.is_some()),
-                "contact pairs require 6-DOF (rotational state + mass) on all bodies"
-            );
-            // Derivative-class thermal (DerivativeFirstOrder /
-            // DerivativeRk4) is not extended to the contact-coupled kernel.
-            // JEOD's `DynamicsIntegrationGroup` handles this case natively,
-            // but our `integrate_bodies_contact_coupled` has no per-stage
-            // SRP/thermal hook yet; opt such bodies into
-            // `ThermalIntegrationOrder::Scheduled` or disable contact pairs.
-            assert!(
-                self.bodies.iter().all(|b| b
-                    .flat_plate_state
-                    .as_ref()
-                    .is_none_or(|fps| fps.stage_inputs.is_none())),
-                "Derivative-class thermal integration is not yet supported with contact pairs; \
-                 use ThermalIntegrationOrder::Scheduled on flat-plate SRP bodies \
-                 when contact pairs are active",
-            );
-            // Contact pair states must share the root inertial frame, since
-            // the coupled contact evaluator uses each body's stage state
-            // directly without any per-step frame transform. `validate()`
-            // catches this at config time; the assert is defense-in-depth
-            // for callers that skip validation.
-            assert!(
-                self.contact_pairs.iter().all(|p| {
-                    let fa = self.bodies[p.body_a].integ_frame_id;
-                    let fb = self.bodies[p.body_b].integ_frame_id;
-                    fa == fb && fa == self.root_frame_id
-                }),
-                "contact pair bodies must share the root inertial integration frame"
-            );
-
-            // `integ_dt` (dynamic timestep) is defined above the gravity
-            // closure; reuse it here for the coupled integrator call.
-
-            // Split disjoint `self` fields up front so we can keep mutable
-            // access to bodies (and to coupled_integ_scratch below) while
-            // borrowing contact pairs immutably — no per-step clone of the
-            // facet/material data.
-            let contact_pairs: &Vec<ContactPairConfig> = &self.contact_pairs;
-            let bodies_mut = &mut self.bodies;
-
-            // Gather per-body immutable data (t_struct_body, mass, constant
-            // forces/torques, gravity_controls) BEFORE the per-body
-            // `iter_mut()` projection below builds the `CoupledBodyInput`
-            // vector. Once that vector is live, holding any shared borrow
-            // into `bodies_mut` (even for a disjoint field on a different
-            // body) would conflict with the &mut field projections, so
-            // snapshot everything up front. Cloning `gravity_controls`
-            // happens here (and only here, since contact_pairs is
-            // non-empty) — the non-contact path above borrows directly.
-            let t_struct_body_vec: Vec<DMat3> =
-                bodies_mut.iter().map(|b| b.t_struct_body).collect();
-            let mass_vec: Vec<MassProperties> = bodies_mut
-                .iter()
-                .map(|b| b.mass.expect("validated"))
-                .collect();
-            let non_grav_non_contact_vec: Vec<DVec3> =
-                bodies_mut.iter().map(|b| b.total_force.force).collect();
-            let non_contact_torque_vec: Vec<DVec3> =
-                bodies_mut.iter().map(|b| b.total_force.torque).collect();
-            let per_body_gravity_controls: Vec<GravityControls<usize>> = bodies_mut
-                .iter()
-                .map(|b| b.gravity_controls.clone())
-                .collect();
-
-            // Build coupled inputs. `CoupledBodyInput` needs separate &mut
-            // borrows for `trans` and `rot.as_mut()` on each `SimBody`. The
-            // borrow checker accepts this when both come out of a single
-            // `iter_mut()` chain — each closure invocation receives a
-            // distinct `&mut SimBody` (the iterator hands them out one at
-            // a time), and disjoint-field projection within that body
-            // produces the two split mutable borrows.
-            //
-            // Earlier revisions used raw pointers in an `unsafe` block to
-            // simulate the same effect, but the immutable-snapshot
-            // pattern above (mass_vec, non_grav_*_vec, …) means no
-            // shared borrow into `bodies_mut` is held while the &muts
-            // are live, so the safe projection compiles cleanly.
-            //
-            // The `expect`s match the validated invariants — the
-            // contact-coupled path requires 6-DOF + 3-component mass on
-            // every body (enforced by `Self::validate`).
-            let mut inputs: Vec<CoupledBodyInput<'_>> = bodies_mut
-                .iter_mut()
-                .enumerate()
-                .map(|(i, body)| CoupledBodyInput {
-                    trans: &mut body.trans,
-                    rot: body
-                        .rot
-                        .as_mut()
-                        .expect("validated: 6-DOF required for contact-coupled path"),
-                    mass: &mass_vec[i],
-                    non_grav_non_contact_force: non_grav_non_contact_vec[i],
-                    non_contact_torque_body: non_contact_torque_vec[i],
-                })
-                .collect();
-
-            integrate_bodies_contact_coupled(
-                &mut inputs,
-                &mut self.coupled_integ_scratch,
-                |body_idx: usize, pos: DVec3, vel: DVec3, time_frac: f64| {
-                    eval_body_gravity(
-                        &per_body_gravity_controls[body_idx],
-                        body_idx,
-                        pos,
-                        vel,
-                        time_frac,
-                    )
-                },
-                |stage_trans: &[TranslationalState],
-                 stage_rot: &[RotationalState],
-                 out: &mut [(DVec3, DVec3)]| {
-                    // Evaluate every registered contact pair at the stage
-                    // states and accumulate force/torque on each body. The
-                    // integrator (`eval_stage` in jeod_sim::integration)
-                    // zeroes `out` before calling us, so this closure just
-                    // accumulates.
-                    for pair in contact_pairs {
-                        if let Some(eval) = evaluate_contact_pair(
-                            &pair.facet_a,
-                            &pair.facet_b,
-                            &stage_trans[pair.body_a],
-                            &stage_trans[pair.body_b],
-                            Some(&stage_rot[pair.body_a]),
-                            Some(&stage_rot[pair.body_b]),
-                            t_struct_body_vec[pair.body_a],
-                            t_struct_body_vec[pair.body_b],
-                            Some(&mass_vec[pair.body_a]),
-                            Some(&mass_vec[pair.body_b]),
-                        ) {
-                            out[pair.body_a].0 += eval.force_on_a;
-                            out[pair.body_b].0 -= eval.force_on_a;
-                            out[pair.body_a].1 += eval.torque_a_body;
-                            out[pair.body_b].1 += eval.torque_b_body;
-                        }
-                    }
-                },
-                integ_dt,
-            );
-        }
-
-        // Sync body positions back to frame tree after integration.
-        for body in &self.bodies {
-            let node = self.frame_tree.get_mut(body.body_frame_id);
-            node.state.trans.position = body.trans.position;
-            node.state.trans.velocity = body.trans.velocity;
-        }
-
-        // ── 8b. Frame switch (body actions) ──
-        // Applied AFTER integration, matching JEOD's pipeline where
-        // DynBodyFrameSwitch is a body action evaluated post-integration.
-        // The body has already been integrated in its current frame for this
-        // step; the switch transforms to the new frame for the NEXT step.
-        // Uses frame tree reparenting for structural correctness.
-        // Use index-based loop to avoid borrow conflict with self.frame_tree.
-        for body_idx in 0..self.bodies.len() {
-            if self.bodies[body_idx].frame_switches.is_empty() {
-                continue;
-            }
-            let mut switch_idx = None;
-            for (idx, sw) in self.bodies[body_idx].frame_switches.iter().enumerate() {
-                if !sw.active {
-                    continue;
-                }
-                let num_sources = self.source_frame_ids.len();
-                let target_fid = self
-                    .source_frame_ids
-                    .get(sw.target_source)
-                    .ok_or(StepError::FrameSwitchTargetMissing {
-                        body_idx,
-                        target_source: sw.target_source,
-                        num_sources,
-                    })?
-                    .inertial;
-                let (target_origin, _) = self.frame_origin(target_fid);
-                let (current_origin, _) = self.frame_origin(self.bodies[body_idx].integ_frame_id);
-                let body_pos_eci = self.bodies[body_idx].trans.position + current_origin;
-                let threshold_sq = sw.switch_distance * sw.switch_distance;
-
-                // JEOD dyn_body_frame_switch.cc:173-182:
-                // OnApproach: compute_position_from(*integ_frame) → distance to target
-                // OnDeparture: state.trans.position magnitude → distance from current origin
-                let triggered = match sw.switch_sense {
-                    SwitchSense::OnApproach => {
-                        (body_pos_eci - target_origin).length_squared() < threshold_sq
-                    }
-                    SwitchSense::OnDeparture => {
-                        self.bodies[body_idx].trans.position.length_squared() > threshold_sq
-                    }
-                };
-                if triggered {
-                    switch_idx = Some(idx);
-                    break;
-                }
-            }
-            if let Some(idx) = switch_idx {
-                let target_source = self.bodies[body_idx].frame_switches[idx].target_source;
-                self.bodies[body_idx].frame_switches[idx].active = false;
-
-                let new_integ_fid = self.source_frame_ids[target_source].inertial; // bounds already checked above
-                let body_fid = self.bodies[body_idx].body_frame_id;
-
-                // Reparent body frame in tree (preserves absolute state).
-                self.frame_tree.reparent(body_fid, new_integ_fid);
-                let new_state = self.frame_tree.get(body_fid).state;
-                self.bodies[body_idx].trans.position = new_state.trans.position;
-                self.bodies[body_idx].trans.velocity = new_state.trans.velocity;
-                self.bodies[body_idx].integ_frame_id = new_integ_fid;
-
-                // Flip gravity controls: target source becomes non-differential
-                // (central body), all others become differential.
-                for ctrl in &mut self.bodies[body_idx].gravity_controls.controls {
-                    ctrl.differential = ctrl.source_name != target_source;
-                }
-            }
-        }
-
-        // ── 9. Derived states ──
-        let gravity_data = &self.gravity_data;
-
-        for body in &mut self.bodies {
-            // Orbital elements
-            if let Some(src_idx) = body.orbital_elements_source {
-                if let Some(mu) = gravity_data.get(src_idx).map(|g| g.source.mu) {
-                    body.orbital_elements = jeod_sim::compute_orbital_elements(
-                        mu,
-                        body.trans.position,
-                        body.trans.velocity,
-                    )
-                    .ok();
-                } else {
-                    body.orbital_elements = None;
-                }
-            }
-
-            // Euler angles
-            if let Some(seq) = body.euler_sequence {
-                if let Some(ref rot) = body.rot {
-                    body.euler_angles = Some(jeod_sim::compute_body_euler_angles(rot, seq));
-                } else {
-                    body.euler_angles = None;
-                }
-            }
-
-            // LVLH frame
-            if body.compute_lvlh {
-                body.lvlh_frame = Some(jeod_sim::compute_body_lvlh_frame(
-                    body.trans.position,
-                    body.trans.velocity,
-                ));
-            }
-
-            // Geodetic state
-            if let Some((src_idx, r_eq, r_pol)) = body.geodetic_planet {
-                let pfix_rot = self
-                    .source_frame_ids
-                    .get(src_idx)
-                    .and_then(|sfids| sfids.pfix)
-                    .map(|pfix_id| self.frame_tree.get(pfix_id).state.rot.t_parent_this);
-                if let Some(t_pfix) = pfix_rot {
-                    body.geodetic_state = Some(jeod_sim::compute_body_geodetic(
-                        body.trans.position,
-                        &t_pfix,
-                        r_eq,
-                        r_pol,
-                    ));
-                } else {
-                    body.geodetic_state = None;
-                }
-            }
-
-            // Solar beta
-            if body.compute_solar_beta {
-                if let Some(sp) = sun_pos {
-                    body.solar_beta = Some(jeod_sim::compute_body_solar_beta(
-                        body.trans.position,
-                        body.trans.velocity,
-                        sp,
-                    ));
-                } else {
-                    body.solar_beta = None;
-                }
-            }
-
-            // Earth lighting
-            if let Some((earth_r, moon_r, sun_r)) = body.earth_lighting_config {
-                if let (Some(sp), Some(mp)) = (sun_pos, moon_pos) {
-                    body.earth_lighting =
-                        Some(jeod_interactions::earth_lighting::compute_earth_lighting(
-                            body.trans.position,
-                            sp,
-                            mp,
-                            sun_r,
-                            earth_r,
-                            moon_r,
-                        ));
-                } else {
-                    body.earth_lighting = None;
-                }
-            }
-        }
-
-        // Advance any free-flying detached subtrees ballistically. This
-        // matches JEOD's behavior for tree roots whose grav_interaction
-        // is empty (the common case for staging — no force applied to
-        // separated stages between detach and reattach). Use the dynamic
-        // timestep `dt * time_scale_factor` so the ballistic propagation
-        // stays consistent with the integrated bodies under time reversal
-        // / scaling (matches `integ_dt` used elsewhere in this step).
-        if !self.detached_subtrees.is_empty() {
-            self.step_detached_subtrees(dt * self.time.time_scale_factor);
-        }
-
-        Ok(())
-    }
-
-    /// Advance the simulation by `n` timesteps.
-    pub fn step_n(&mut self, n: usize) -> Result<(), StepError> {
-        for _ in 0..n {
-            self.step()?;
-        }
-        Ok(())
-    }
-
-    /// Advance the simulation until `target_time` (in simulation seconds).
-    ///
-    /// Steps at `self.dt` until the remaining time is less than `dt`,
-    /// then takes a final fractional step if the remainder exceeds 1 ms.
-    pub fn step_until(&mut self, target_time: f64) -> Result<(), StepError> {
-        while self.time.simtime + self.dt <= target_time + 0.001 {
-            self.step()?;
-        }
-        let remainder = target_time - self.time.simtime;
-        if remainder > 0.001 {
-            // Fractional steps corrupt multi-step history arrays (GJ's
-            // Störmer-Cowell coefficients and delinv accumulators, ABM4's
-            // Adams history) — both methods assume constant dt.
-            let has_multistep = self.bodies.iter().any(|b| {
-                matches!(
-                    b.integrator,
-                    jeod_dynamics::IntegratorType::GaussJackson(..)
-                        | jeod_dynamics::IntegratorType::Abm4
-                )
-            });
-            assert!(
-                !has_multistep,
-                "step_until() would take a fractional step ({remainder:.6}s vs dt={:.6}s). \
-                 Multi-step integrators (GaussJackson, ABM4) require constant dt. \
-                 Ensure target_time is an integer multiple of dt.",
-                self.dt
-            );
-            self.step_internal(remainder)?;
-        }
-        Ok(())
-    }
-
-    // JEOD_INV: DS.01 — derived state config immutable after init; read-only access only
-    /// Get the current output state of a body by index.
-    ///
-    /// Returns a [`VehicleOutput`] containing the current integrated state
-    /// plus any derived states that were configured.
-    pub fn body(&self, idx: usize) -> VehicleOutput {
-        self.bodies[idx].output()
-    }
-
-    /// Adjust an integrated body's `trans` from a `core_body` inertial
-    /// state to the corresponding `composite_body` inertial state,
-    /// using the current mass tree's `core_wrt_composite` offset.
-    ///
-    /// JEOD's integration variable is `composite_body`; at init time,
-    /// however, callers often have JEOD-published values that were
-    /// logged from `core_body` (as in our `tier3_sim_apollo_trajectory`
-    /// reference CSV). Use this once after the mass tree has reached
-    /// its initial topology to flip the interpretation.
-    ///
-    /// `body.rot` is unchanged — composite and core share body axes
-    /// (see `body_core_inertial` for the full convention).
-    ///
-    /// # Panics
-    /// Panics if the body is not registered in the mass tree, or has
-    /// no rotational state.
-    pub fn convert_body_trans_core_to_composite(&mut self, idx: usize) {
-        let cw_inertial;
-        let dvel_inertial;
-        {
-            let body = &self.bodies[idx];
-            let mass_body_id = body.mass_body_id.expect(
-                "convert_body_trans_core_to_composite: body is not registered in the mass tree",
-            );
-            let tree = self
-                .mass_tree
-                .as_ref()
-                .expect("convert_body_trans_core_to_composite: no mass tree configured");
-            let node = tree.get(mass_body_id);
-            let cw_struct = node.core_wrt_composite.position;
-            let t_struct_to_body = node.composite_properties.t_parent_this;
-            let cw_body = t_struct_to_body * cw_struct;
-            let body_rot = body
-                .rot
-                .expect("convert_body_trans_core_to_composite: 6-DOF body required");
-            let t_inertial_to_body = body_rot.quaternion.left_quat_to_transformation();
-            let t_body_to_inertial = t_inertial_to_body.transpose();
-            cw_inertial = t_body_to_inertial * cw_body;
-            dvel_inertial = t_body_to_inertial * body_rot.ang_vel_body.cross(cw_body);
-        }
-        // composite = core − cw_inertial; subtract the rigid-body
-        // ω × r contribution on velocity.
-        self.bodies[idx].trans.position -= cw_inertial;
-        self.bodies[idx].trans.velocity -= dvel_inertial;
-    }
-
-    /// Derive the integrated body's `core_body` inertial position +
-    /// velocity from its `composite_body` integration state and the
-    /// current mass tree.
-    ///
-    /// JEOD integrates `composite_body` (matching
-    /// `DynamicsIntegrationGroup::gravitation()` and
-    /// `DynBody::trans_integ()`), so during stages 1–6 of the JEOD
-    /// integration loop `body.trans` represents the composite. The
-    /// `core_body` frame is the per-body CoM, derived via the mass
-    /// tree's `core_wrt_composite` offset rotated to inertial.
-    ///
-    /// In our mass tree `composite_properties.t_parent_this` is left
-    /// equal to `core_properties.t_parent_this` (set per-body at init,
-    /// not re-derived for the merged composite), so composite and core
-    /// share body axes — only position and velocity differ. Returns
-    /// `(position, velocity)` in the body's integration-frame inertial
-    /// coordinates (same frame as `body.trans`).
-    ///
-    /// # Panics
-    /// Panics if the body is not registered in the mass tree, or if
-    /// the body has no rotational state (6-DOF required to derive the
-    /// kinematic offset).
-    pub fn body_core_inertial(&self, idx: usize) -> (DVec3, DVec3) {
-        let body = &self.bodies[idx];
-        let mass_body_id = body
-            .mass_body_id
-            .expect("body_core_inertial: body is not registered in the mass tree");
-        let tree = self
-            .mass_tree
-            .as_ref()
-            .expect("body_core_inertial: no mass tree configured");
-        let node = tree.get(mass_body_id);
-        let core_wrt_composite_struct = node.core_wrt_composite.position;
-        // Struct → composite-body rotation (composite shares core's body axes).
-        let t_struct_to_body = node.composite_properties.t_parent_this;
-        let cw_body = t_struct_to_body * core_wrt_composite_struct;
-        // Body → inertial via T_inertial_to_body⁻¹ = T_inertial_to_body.transpose().
-        let body_rot = body.rot.expect("body_core_inertial: 6-DOF body required");
-        let t_inertial_to_body = body_rot.quaternion.left_quat_to_transformation();
-        let t_body_to_inertial = t_inertial_to_body.transpose();
-        let cw_inertial = t_body_to_inertial * cw_body;
-        let core_position = body.trans.position + cw_inertial;
-        // v_core = v_composite + ω × r (in inertial frame). ω in body
-        // frame is body.rot.ang_vel_body; rotate the cross-product to
-        // inertial.
-        let omega_body = body_rot.ang_vel_body;
-        let core_velocity = body.trans.velocity + t_body_to_inertial * omega_body.cross(cw_body);
-        (core_position, core_velocity)
-    }
-
-    /// Resolve the `composite_body` inertial state of any mass-tree body,
-    /// regardless of whether it is the integrated body, attached as a
-    /// child of the integrated body, or the root of a detached subtree.
-    ///
-    /// Walks to the tree root, reads the root's inertial composite state
-    /// from either the integrated-body slot ([`Self::body`]) or the
-    /// [`Self::detached_subtrees`] map (when the root is detached), then
-    /// chain-walks down to `target_id` using the same body-aware step as
-    /// [`Self::detach_subtree`] (with [`jeod_dynamics::propagate_forward`]
-    /// at each level).
-    ///
-    /// This mirrors what JEOD's truth recorder logs for `lm_dyn.composite_body`
-    /// regardless of the LM's current attach state — the
-    /// `tier3_sim_apollo_lm_state_vs_truth` diagnostic compares against
-    /// `apollo_attach_truth.csv` rows produced by that same recorder.
-    ///
-    /// # Panics
-    /// Panics if no mass tree is configured, `target_id` is not in the
-    /// tree, the root has no integrated body and no detached-subtree
-    /// entry, or the integrated root is missing rotational state.
-    pub fn subtree_composite_inertial(&self, target_id: MassBodyId) -> RefFrameState {
-        let tree = self
-            .mass_tree
-            .as_ref()
-            .expect("subtree_composite_inertial: no mass tree configured");
-        // Walk to root.
-        let mut root_id = target_id;
-        while let Some(p) = tree.parent(root_id) {
-            root_id = p;
-        }
-
-        // Resolve the root's inertial composite_body state.
-        let integrated_idx = self
-            .bodies
-            .iter()
-            .position(|b| b.mass_body_id == Some(root_id));
-        let root_state: RefFrameState = if let Some(idx) = integrated_idx {
-            let body = &self.bodies[idx];
-            let body_rot = body.rot.expect(
-                "subtree_composite_inertial: integrated root has no rotational state \
-                 (6-DOF required)",
-            );
-            RefFrameState {
-                trans: RefFrameTrans {
-                    position: body.trans.position,
-                    velocity: body.trans.velocity,
-                },
-                rot: RefFrameRot {
-                    q_parent_this: body_rot.quaternion,
-                    t_parent_this: body_rot.quaternion.left_quat_to_transformation(),
-                    ang_vel_this: body_rot.ang_vel_body,
-                },
-            }
-        } else if let Some(detached) = self.detached_subtrees.get(&root_id) {
-            detached.to_ref_frame_state()
-        } else {
-            panic!(
-                "subtree_composite_inertial: tree root {root_id:?} for target \
-                 {target_id:?} has no integrated body and no detached-subtree entry"
-            );
-        };
-
-        // Build the chain root → target and walk down with the body-aware step.
-        let mut chain = Vec::<MassBodyId>::new();
-        let mut cur = target_id;
-        while cur != root_id {
-            chain.push(cur);
-            cur = tree.parent(cur).expect(
-                "subtree_composite_inertial: chain walk hit a parentless intermediate \
-                 (target lost its parent during traversal)",
-            );
-        }
-        chain.reverse();
-
-        let mut current_state = root_state;
-        let mut current_node_id = root_id;
-        for next_id in chain {
-            let next_node = tree.get(next_id);
-            let current_node = tree.get(current_node_id);
-            let t_current_struct_to_body = current_node.composite_properties.t_parent_this;
-            let t_next_struct_to_body = next_node.composite_properties.t_parent_this;
-            let offset_struct =
-                next_node.composite_wrt_pstr.position - current_node.composite_properties.position;
-            let offset_in_current_body = t_current_struct_to_body * offset_struct;
-            let t_current_body_to_next_body = t_next_struct_to_body
-                * next_node.structure_point.t_parent_this
-                * t_current_struct_to_body.transpose();
-            let rel = MassPointState {
-                position: offset_in_current_body,
-                t_parent_this: t_current_body_to_next_body,
-            };
-            current_state = jeod_dynamics::propagate_forward(&current_state, &rel);
-            current_node_id = next_id;
-        }
-        current_state
-    }
-
-    /// Read the current per-plate temperatures (K) for a body's flat-plate
-    /// SRP configuration, or `None` if the body has no flat-plate SRP.
-    ///
-    /// Useful for unit tests and data recording; returns a reference so no
-    /// allocation is needed for the common steady-state case.
-    pub fn srp_plate_temperatures(&self, body_idx: usize) -> Option<&[f64]> {
-        self.bodies[body_idx]
-            .flat_plate_state
-            .as_ref()
-            .map(|fps| fps.temperatures.as_slice())
-    }
-
-    /// Set the externally applied force (inertial frame, N) for a body.
-    ///
-    /// Added to `total_force.force` each step after force collection.
-    pub fn set_body_external_force(&mut self, idx: usize, force: DVec3) {
-        self.bodies[idx].external_force = force;
-    }
-
-    /// Set the externally applied torque (body frame, N*m) for a body.
-    ///
-    /// Added to `total_force.torque` each step after force collection.
-    pub fn set_body_external_torque(&mut self, idx: usize, torque: DVec3) {
-        self.bodies[idx].external_torque = torque;
-    }
-
-    /// Set a body's translational position (inertial frame, m).
-    ///
-    /// Used for prescribed-motion tests where position is set externally
-    /// at each timestep (e.g., SIM_2A_SHADOW_CALC).
-    pub fn set_body_position(&mut self, idx: usize, position: DVec3) {
-        self.bodies[idx].trans.position = position;
-        let fid = self.bodies[idx].body_frame_id;
-        self.frame_tree.get_mut(fid).state.trans.position = position;
-    }
-
-    /// Set a body's translational velocity (inertial frame, m/s).
-    ///
-    /// Used for impulsive maneuvers (e.g., Apollo TLI delta-V).
-    pub fn set_body_velocity(&mut self, idx: usize, velocity: DVec3) {
-        self.bodies[idx].trans.velocity = velocity;
-        let fid = self.bodies[idx].body_frame_id;
-        self.frame_tree.get_mut(fid).state.trans.velocity = velocity;
-    }
-
-    /// Replace a body's mass properties.
-    ///
-    /// Used for discrete mass changes (e.g., post-burn fuel consumption,
-    /// stage separation). Recomputes `inverse_mass` and `inverse_inertia`.
-    ///
-    /// **Warning:** If the body is registered in the mass tree, calling this
-    /// method will desynchronize the body's mass from the tree's copy. Use
-    /// [`sync_body_mass_from_tree`](Self::sync_body_mass_from_tree) instead
-    /// when the mass tree has been modified via `attach`/`detach`.
-    pub fn set_body_mass(&mut self, idx: usize, mut mass: MassProperties) {
-        mass.dirty = true;
-        mass.recompute_derived();
-        self.bodies[idx].mass = Some(mass);
-    }
-
-    /// Sync a body's mass properties from the mass tree's composite.
-    ///
-    /// After modifying the mass tree via `attach`/`detach`, call this to
-    /// update the body's mass from the tree's composite properties.
-    ///
-    /// # Panics
-    /// Panics if the body is not registered in the mass tree.
-    pub fn sync_body_mass_from_tree(&mut self, idx: usize) {
-        let id = self.bodies[idx]
-            .mass_body_id
-            .expect("sync_body_mass_from_tree requires body registered in mass tree");
-        let tree = self
-            .mass_tree
-            .as_ref()
-            .expect("sync_body_mass_from_tree requires a mass tree");
-        let mut composite = tree.get(id).composite_properties;
-        composite.dirty = true;
-        composite.recompute_derived();
-        self.bodies[idx].mass = Some(composite);
-    }
-
-    /// Register a body in the simulation's mass tree.
-    ///
-    /// Creates (or reuses) a `MassTree` and adds the body's mass as a node.
-    /// Returns the `MassBodyId` for use with [`attach`](Self::attach) and
-    /// [`detach`](Self::detach). The body's `mass` field must be `Some`.
-    ///
-    /// # Panics
-    /// Panics if the body has no mass properties.
-    pub fn add_body_to_tree(
-        &mut self,
-        body_idx: usize,
-        name: impl Into<String>,
-    ) -> jeod_dynamics::MassBodyId {
-        let mass = self.bodies[body_idx]
-            .mass
-            .expect("add_body_to_tree requires mass properties");
-        let tree = self
-            .mass_tree
-            .get_or_insert_with(jeod_dynamics::MassTree::new);
-        let id = tree.add_body(name.into(), mass);
-        self.bodies[body_idx].mass_body_id = Some(id);
-        id
-    }
-
-    /// Attach a child body to a parent body in the mass tree.
-    ///
-    /// Both bodies must have been registered via [`add_body_to_tree`](Self::add_body_to_tree).
-    /// After attachment, the parent's composite mass properties are updated
-    /// automatically. The parent body's `mass` is synced from the tree.
-    ///
-    /// # Panics
-    /// Panics if either body is not in the tree, or if the child already has a parent.
-    pub fn attach(
-        &mut self,
-        child_idx: usize,
-        parent_idx: usize,
-        offset: DVec3,
-        t_parent_child: DMat3,
-    ) {
-        let child_id = self.bodies[child_idx]
-            .mass_body_id
-            .expect("attach: child body not in mass tree");
-        let parent_id = self.bodies[parent_idx]
-            .mass_body_id
-            .expect("attach: parent body not in mass tree");
-        let tree = self.mass_tree.as_mut().expect("attach: no mass tree");
-        tree.attach(child_id, parent_id, offset, t_parent_child);
-        // Sync parent's composite mass from tree
-        self.bodies[parent_idx].mass = Some(tree.get(parent_id).composite_properties);
-    }
-
-    /// Detach a child body from its parent in the mass tree.
-    ///
-    /// After detachment, both the former parent's and the child's mass
-    /// properties are updated from the tree's recomputed composites.
-    ///
-    /// # Panics
-    /// Panics if the body is not in the tree or has no parent.
-    pub fn detach(&mut self, child_idx: usize) {
-        let child_id = self.bodies[child_idx]
-            .mass_body_id
-            .expect("detach: child body not in mass tree");
-        let tree = self.mass_tree.as_mut().expect("detach: no mass tree");
-        let parent_id = tree
-            .parent(child_id)
-            .expect("detach: child body has no parent in tree");
-        tree.detach(child_id);
-        // Sync both bodies' mass from tree
-        self.bodies[child_idx].mass = Some(tree.get(child_id).composite_properties);
-        // Find parent body index and sync
-        for body in &mut self.bodies {
-            if body.mass_body_id == Some(parent_id) {
-                body.mass = Some(tree.get(parent_id).composite_properties);
-                break;
-            }
-        }
-    }
-
-    /// Detach a tree-only subtree from its parent in the mass tree,
-    /// capturing the subtree's composite-body inertial state at the
-    /// moment of separation.
-    ///
-    /// The parent of the subtree may be either the integrated body's
-    /// own mass-tree node or another *already-detached* subtree (whose
-    /// state lives in [`Simulation::detached_subtrees`]). The method
-    /// locates the parent automatically by walking up the tree from
-    /// `subtree_root_id` to its root.
-    ///
-    /// `integrated_body_idx` is consulted only to identify the
-    /// Simulation body whose mass-tree id matches the parent's tree
-    /// root (when the parent is the integrated body).
-    /// `subtree_root_id` is the [`MassBodyId`] being detached.
-    ///
-    /// At detach time the subtree shared the parent's rigid motion, so
-    /// its composite-CoM inertial state is computed from the parent's
-    /// state plus the offset between the two composites (drawn from
-    /// the mass tree). The state is then stored on
-    /// [`Simulation::detached_subtrees`] and propagated each step
-    /// until [`attach_subtree_aligned`](Self::attach_subtree_aligned)
-    /// re-attaches the subtree.
-    ///
-    /// If the parent is a detached subtree, its tracked composite state
-    /// is updated to reflect the loss of mass (the parent's
-    /// composite-CoM in inertial shifts when its mass distribution
-    /// changes — even though the underlying structure point doesn't
-    /// move). If the parent is the integrated body, `body.trans`
-    /// (the integrated `composite_body` inertial state, post-`bd279c2`)
-    /// is shifted by the inertial-frame composite-CoM delta so it
-    /// continues to track the new (smaller) composite, and the body's
-    /// mass is re-synced from the recomputed composite_properties.
-    ///
-    /// # Panics
-    /// Panics if no mass tree is configured, the subtree id is not in
-    /// the tree, the subtree has no parent, the parent's root has no
-    /// tracked state, or a subtree with the same id is already in the
-    /// detached map.
-    pub fn detach_subtree(&mut self, integrated_body_idx: usize, subtree_root_id: MassBodyId) {
-        let tree = self
-            .mass_tree
-            .as_ref()
-            .expect("detach_subtree: no mass tree configured");
-        // Walk up to find the root of subtree_root_id's current tree.
-        tree.parent(subtree_root_id)
-            .expect("detach_subtree: subtree has no parent in tree");
-        let mut tree_root_id = subtree_root_id;
-        while let Some(p) = tree.parent(tree_root_id) {
-            tree_root_id = p;
-        }
-
-        // The parent's pre-detach composite-CoM offset in its own struct frame.
-        let parent_pre_composite_props = tree.get(tree_root_id).composite_properties;
-
-        // Determine where the parent's state lives — either an integrated
-        // Simulation body or a detached subtree.
-        let integrated_mass_body_id = self.bodies[integrated_body_idx].mass_body_id;
-        let parent_is_integrated = integrated_mass_body_id == Some(tree_root_id);
-
-        // Pre-detach inertial composite_body state of the parent.
-        // body.trans / body.rot represent the integrated body's
-        // composite_body inertial state (matching JEOD's integration
-        // variable; see `attach_subtree_aligned` and the "Integration
-        // target" note on `Simulation`).
-        let parent_composite_state: RefFrameState = if parent_is_integrated {
-            let body_trans = self.bodies[integrated_body_idx].trans;
-            let body_rot = self.bodies[integrated_body_idx]
-                .rot
-                .expect("detach_subtree: 6-DOF integrated body required");
-            RefFrameState {
-                trans: RefFrameTrans {
-                    position: body_trans.position,
-                    velocity: body_trans.velocity,
-                },
-                rot: RefFrameRot {
-                    q_parent_this: body_rot.quaternion,
-                    t_parent_this: body_rot.quaternion.left_quat_to_transformation(),
-                    ang_vel_this: body_rot.ang_vel_body,
-                },
-            }
-        } else {
-            // Parent is itself a detached subtree.
-            let detached = self
-                .detached_subtrees
-                .get(&tree_root_id)
-                .unwrap_or_else(|| {
-                    panic!(
-                        "detach_subtree: parent tree-root {tree_root_id:?} of \
-                         subtree {subtree_root_id:?} has no tracked state — \
-                         did you forget to call detach_subtree on it first?"
-                    )
-                });
-            detached.to_ref_frame_state()
-        };
-
-        // Walk down the tree from the root to the subtree, applying
-        // propagate_forward at each level. This handles arbitrary tree
-        // depth (e.g. cm → sm → s3 → lm where the subtree being
-        // detached is several levels below the root). Each level uses
-        // the immediate-parent-struct-frame `composite_wrt_pstr` from
-        // the mass tree.
-        let mut chain = Vec::<MassBodyId>::new();
-        let mut current_id = subtree_root_id;
-        while current_id != tree_root_id {
-            chain.push(current_id);
-            current_id = tree
-                .parent(current_id)
-                .expect("detach_subtree: chain walk hit a parentless intermediate");
-        }
-        chain.reverse(); // tree_root → ... → subtree
-        let mut current_state = parent_composite_state;
-        let mut current_node_id = tree_root_id;
-        for next_id in chain {
-            let next_node = tree.get(next_id);
-            let current_node = tree.get(current_node_id);
-            // Body-aware step:
-            //   offset_in_current_body = T_current_struct_to_body
-            //                          · (next.composite_wrt_pstr.position
-            //                             − current.composite_properties.position)
-            //   T_current_body_to_next_body = T_next_struct_to_body
-            //                               · next.structure_point.t_parent_this
-            //                               · T_current_body_to_struct
-            // For axis-aligned bodies (struct == body) this collapses to
-            // the simple struct-frame difference and the bare structure_point
-            // rotation.
-            let t_current_struct_to_body = current_node.composite_properties.t_parent_this;
-            let t_next_struct_to_body = next_node.composite_properties.t_parent_this;
-            let offset_struct =
-                next_node.composite_wrt_pstr.position - current_node.composite_properties.position;
-            let offset_in_current_body = t_current_struct_to_body * offset_struct;
-            let t_current_body_to_next_body = t_next_struct_to_body
-                * next_node.structure_point.t_parent_this
-                * t_current_struct_to_body.transpose();
-            let rel = MassPointState {
-                position: offset_in_current_body,
-                t_parent_this: t_current_body_to_next_body,
-            };
-            current_state = jeod_dynamics::propagate_forward(&current_state, &rel);
-            current_node_id = next_id;
-        }
-        let subtree_state = current_state;
-
-        // Apply the topology change — this also recomputes parent's
-        // composite_properties (now without the subtree).
-        let tree = self.mass_tree.as_mut().unwrap();
-        tree.detach(subtree_root_id);
-        let parent_post_composite_props = tree.get(tree_root_id).composite_properties;
-
-        if parent_is_integrated {
-            // body.trans/body.rot are the integrated composite_body
-            // state. JEOD's detach handler preserves core_body
-            // (Pos_Vel_Att_Rate source = core_body) and rederives the
-            // post-detach composite from it. The composite-CoM offset
-            // in the parent's struct frame shifts when the subtree
-            // leaves, so the inertial composite_body position +
-            // velocity must shift by the corresponding kinematic
-            // offset. Rotation/ang_vel are unchanged because
-            // composite_properties.t_parent_this == core_properties
-            // .t_parent_this throughout (see mass tree recompute).
-            let cm_delta_struct =
-                parent_post_composite_props.position - parent_pre_composite_props.position;
-            // composite_properties.t_parent_this is the struct→body
-            // rotation; compose with body.rot's t_parent_this to map
-            // struct → inertial.
-            let t_struct_to_body = parent_pre_composite_props.t_parent_this;
-            let cm_delta_body = t_struct_to_body * cm_delta_struct;
-            let t_inertial_to_body = parent_composite_state.rot.t_parent_this;
-            let cm_delta_inertial = t_inertial_to_body.transpose() * cm_delta_body;
-            // Velocity offset from rigid-body rotation: ω × Δr in body
-            // frame, then rotated to inertial.
-            let omega_body = parent_composite_state.rot.ang_vel_this;
-            let dvel_inertial = t_inertial_to_body.transpose() * omega_body.cross(cm_delta_body);
-            self.bodies[integrated_body_idx].trans = TranslationalState {
-                position: parent_composite_state.trans.position + cm_delta_inertial,
-                velocity: parent_composite_state.trans.velocity + dvel_inertial,
-            };
-            // body.rot unchanged — composite/core share body axes.
-            self.bodies[integrated_body_idx].mass = Some(parent_post_composite_props);
-        } else {
-            // Parent is a detached subtree — update its tracked
-            // composite-body state to reflect the new (smaller) composite.
-            // The parent's struct origin in inertial is unchanged (rigid
-            // body); only the composite-CoM has moved within the struct
-            // frame. Convert the struct-frame CoM-delta to inertial via
-            //
-            //   T_inertial_to_struct = T_struct_to_body^T · T_inertial_to_body
-            //
-            // (matching `jeod_dynamics::compute_t_inertial_struct`). The
-            // earlier form `T_struct_to_body * T_inertial_to_body` was
-            // only correct when `T_struct_to_body` is symmetric (identity
-            // or yaw_180) and silently produced wrong CoM-shift directions
-            // for non-symmetric mass-tree orientations.
-            let cm_delta_struct =
-                parent_post_composite_props.position - parent_pre_composite_props.position;
-            let t_struct_to_body = parent_pre_composite_props.t_parent_this;
-            let t_inertial_struct = jeod_dynamics::compute_t_inertial_struct(
-                &t_struct_to_body,
-                &parent_composite_state.rot.t_parent_this,
-            );
-            let r_struct_to_inertial = t_inertial_struct.transpose();
-            let cm_delta_inertial = r_struct_to_inertial * cm_delta_struct;
-            // Velocity contribution from rotation: ω × Δr expressed in
-            // body frame, then rotated to inertial.
-            let w_body = parent_composite_state.rot.ang_vel_this;
-            let cm_delta_body = t_struct_to_body * cm_delta_struct;
-            let dvel_inertial =
-                parent_composite_state.rot.t_parent_this.transpose() * w_body.cross(cm_delta_body);
-            let updated = DetachedSubtreeState {
-                composite_position: parent_composite_state.trans.position + cm_delta_inertial,
-                composite_velocity: parent_composite_state.trans.velocity + dvel_inertial,
-                composite_attitude: parent_composite_state.rot.q_parent_this,
-                composite_ang_vel_body: parent_composite_state.rot.ang_vel_this,
-            };
-            self.detached_subtrees.insert(tree_root_id, updated);
-        }
-
-        if std::env::var("APOLLO_TRACE").is_ok() {
-            eprintln!(
-                "DETACH: subtree {subtree_root_id:?} state stored:\n  pos={:?}\n  vel={:?}\n  ω={:?}",
-                subtree_state.trans.position,
-                subtree_state.trans.velocity,
-                subtree_state.rot.ang_vel_this
-            );
-        }
-
-        // Insert the new subtree's state into the detached map.
-        let prior = self.detached_subtrees.insert(
-            subtree_root_id,
-            DetachedSubtreeState::from_ref_frame_state(&subtree_state),
-        );
-        assert!(
-            prior.is_none(),
-            "detach_subtree: subtree {subtree_root_id:?} was already in detached_subtrees \
-             — call attach_subtree_aligned first or use a fresh subtree id"
-        );
-    }
-
-    /// Re-attach a previously-detached subtree to the integrated body's
-    /// mass tree using named mass points (matching JEOD's
-    /// `attach_aligned`), then update the integrated body's state via
-    /// JEOD's [`combine_states_at_attach`] momentum-conservation
-    /// algorithm.
-    ///
-    /// The integrated body's `body.trans` / `body.rot` represent the
-    /// *composite_body* inertial state of the whole mass tree rooted
-    /// at the integrated body — i.e. the integration variable that
-    /// JEOD's `DynamicsIntegrationGroup::gravitation()` evaluates
-    /// gravity at and that `DynBody::trans_integ()` integrates. The
-    /// subtree state from [`Simulation::detached_subtrees`] is the
-    /// subtree's composite_body frame. After the algorithm runs, the
-    /// integrated body's `trans` / `rot` are set to the new combined
-    /// composite_body inertial state.
-    ///
-    /// To compare against JEOD's logged core_body, derive core via
-    /// [`Simulation::body_core_inertial`].
-    ///
-    /// # Panics
-    /// Panics if the integrated body has no rotational state, no mass
-    /// tree is configured, the parent or subtree id is not in the tree,
-    /// either named mass point is missing on its body, or the subtree
-    /// is not in [`Self::detached_subtrees`].
-    pub fn attach_subtree_aligned(
-        &mut self,
-        integrated_body_idx: usize,
-        subtree_root_id: MassBodyId,
-        subtree_point: &str,
-        parent_id: MassBodyId,
-        parent_point: &str,
-    ) {
-        let tree = self
-            .mass_tree
-            .as_ref()
-            .expect("attach_subtree_aligned: no mass tree configured");
-        let integrated_mass_body_id = self.bodies[integrated_body_idx]
-            .mass_body_id
-            .expect("attach_subtree_aligned: integrated body not registered in mass tree");
-
-        // Read pre-attach composite mass props of the integrated body
-        // (= the whole pre-attach tree without the subtree).
-        let parent_pre_composite_props = tree.get(integrated_mass_body_id).composite_properties;
-        let orig_parent_cm_struct = parent_pre_composite_props.position;
-        let core_wrt_composite_pre = tree.get(integrated_mass_body_id).core_wrt_composite;
-        // Pre-attach subtree composite mass props.
-        let subtree_composite_props = tree.get(subtree_root_id).composite_properties;
-
-        // Read the integrated body's pre-attach composite_body state
-        // directly from body.trans/body.rot (post-refactor convention).
-        let body_trans = self.bodies[integrated_body_idx].trans;
-        let body_rot = self.bodies[integrated_body_idx]
-            .rot
-            .expect("attach_subtree_aligned: 6-DOF body required");
-        let parent_composite_pre = RefFrameState {
-            trans: RefFrameTrans {
-                position: body_trans.position,
-                velocity: body_trans.velocity,
-            },
-            rot: RefFrameRot {
-                q_parent_this: body_rot.quaternion,
-                t_parent_this: body_rot.quaternion.left_quat_to_transformation(),
-                ang_vel_this: body_rot.ang_vel_body,
-            },
-        };
-        let _ = core_wrt_composite_pre; // unused under composite-body convention
-
-        // Borrow the subtree's free-flight composite state (don't remove
-        // yet — we want the entry to survive if any of the operations
-        // below panic, so the caller can retry / recover instead of
-        // silently losing state).
-        let subtree_state = *self
-            .detached_subtrees
-            .get(&subtree_root_id)
-            .unwrap_or_else(|| {
-                panic!(
-                    "attach_subtree_aligned: subtree {subtree_root_id:?} is not in \
-                     detached_subtrees — call detach_subtree first or pre-populate"
-                )
-            });
-        let child_composite = subtree_state.to_ref_frame_state();
-
-        // Apply the topology change (also recomputes composite props).
-        let tree_mut = self.mass_tree.as_mut().unwrap();
-        tree_mut.attach_aligned(subtree_root_id, subtree_point, parent_id, parent_point);
-        // Read post-attach composite props.
-        let combined_composite_props = tree_mut.get(integrated_mass_body_id).composite_properties;
-
-        // The combine algorithm needs `T_inertial_to_struct`, which by
-        // the standard frame-chain rule is
-        //
-        //   T_inertial_to_struct = T_struct_to_body^T · T_inertial_to_body
-        //
-        // (matching `jeod_dynamics::compute_t_inertial_struct` and JEOD's
-        // `dyn_body_collect.cc` lines 219-221). composite_properties
-        // .t_parent_this is the struct→body rotation. The earlier form
-        // `T_struct_to_body * T_inertial_to_body` was only valid for
-        // symmetric struct-to-body rotations (identity, yaw_180 — the
-        // ones Apollo happens to use); non-symmetric vehicle orientations
-        // would silently get a wrong torque arm in the combine algorithm.
-        let t_struct_to_body = parent_pre_composite_props.t_parent_this;
-        let parent_t_inertial_struct = jeod_dynamics::compute_t_inertial_struct(
-            &t_struct_to_body,
-            &parent_composite_pre.rot.t_parent_this,
-        );
-
-        // APOLLO_TRACE diagnostic: dump every input to combine_states_at_attach
-        // so we can diff against JEOD ground truth. Gated by env var so the
-        // regular test path is unaffected. (See #248 attach-bug investigation.)
-        if std::env::var("APOLLO_TRACE").is_ok() {
-            eprintln!("=== ATTACH TRACE (integrated body {integrated_body_idx} → subtree {subtree_root_id:?}) ===");
-            eprintln!("  PARENT COMPOSITE (= our body.trans/body.rot):");
-            eprintln!(
-                "    pos    = [{:.10e} {:.10e} {:.10e}]",
-                parent_composite_pre.trans.position.x,
-                parent_composite_pre.trans.position.y,
-                parent_composite_pre.trans.position.z
-            );
-            eprintln!(
-                "    vel    = [{:.10e} {:.10e} {:.10e}]",
-                parent_composite_pre.trans.velocity.x,
-                parent_composite_pre.trans.velocity.y,
-                parent_composite_pre.trans.velocity.z
-            );
-            eprintln!(
-                "    q      = [{:.10e} {:.10e} {:.10e} {:.10e}]",
-                parent_composite_pre.rot.q_parent_this.scalar(),
-                parent_composite_pre.rot.q_parent_this.vector().x,
-                parent_composite_pre.rot.q_parent_this.vector().y,
-                parent_composite_pre.rot.q_parent_this.vector().z
-            );
-            eprintln!(
-                "    ω_body = [{:.10e} {:.10e} {:.10e}]",
-                parent_composite_pre.rot.ang_vel_this.x,
-                parent_composite_pre.rot.ang_vel_this.y,
-                parent_composite_pre.rot.ang_vel_this.z
-            );
-            eprintln!("  CHILD COMPOSITE (= detached_subtrees[{subtree_root_id:?}]):");
-            eprintln!(
-                "    pos    = [{:.10e} {:.10e} {:.10e}]",
-                child_composite.trans.position.x,
-                child_composite.trans.position.y,
-                child_composite.trans.position.z
-            );
-            eprintln!(
-                "    vel    = [{:.10e} {:.10e} {:.10e}]",
-                child_composite.trans.velocity.x,
-                child_composite.trans.velocity.y,
-                child_composite.trans.velocity.z
-            );
-            eprintln!(
-                "    q      = [{:.10e} {:.10e} {:.10e} {:.10e}]",
-                child_composite.rot.q_parent_this.scalar(),
-                child_composite.rot.q_parent_this.vector().x,
-                child_composite.rot.q_parent_this.vector().y,
-                child_composite.rot.q_parent_this.vector().z
-            );
-            eprintln!(
-                "    ω_body = [{:.10e} {:.10e} {:.10e}]",
-                child_composite.rot.ang_vel_this.x,
-                child_composite.rot.ang_vel_this.y,
-                child_composite.rot.ang_vel_this.z
-            );
-            eprintln!("  PARENT MASS (pre-attach):");
-            eprintln!("    mass={:.10e}", parent_pre_composite_props.mass);
-            eprintln!(
-                "    pos_struct=[{:.10e} {:.10e} {:.10e}]",
-                parent_pre_composite_props.position.x,
-                parent_pre_composite_props.position.y,
-                parent_pre_composite_props.position.z
-            );
-            eprintln!(
-                "    inertia.diag=[{:.10e} {:.10e} {:.10e}]",
-                parent_pre_composite_props.inertia.x_axis.x,
-                parent_pre_composite_props.inertia.y_axis.y,
-                parent_pre_composite_props.inertia.z_axis.z
-            );
-            eprintln!("  CHILD MASS (pre-attach):");
-            eprintln!("    mass={:.10e}", subtree_composite_props.mass);
-            eprintln!(
-                "    pos_struct=[{:.10e} {:.10e} {:.10e}]",
-                subtree_composite_props.position.x,
-                subtree_composite_props.position.y,
-                subtree_composite_props.position.z
-            );
-            eprintln!(
-                "    inertia.diag=[{:.10e} {:.10e} {:.10e}]",
-                subtree_composite_props.inertia.x_axis.x,
-                subtree_composite_props.inertia.y_axis.y,
-                subtree_composite_props.inertia.z_axis.z
-            );
-            eprintln!("  COMBINED MASS (post-attach):");
-            eprintln!("    mass={:.10e}", combined_composite_props.mass);
-            eprintln!(
-                "    pos_struct=[{:.10e} {:.10e} {:.10e}]",
-                combined_composite_props.position.x,
-                combined_composite_props.position.y,
-                combined_composite_props.position.z
-            );
-            eprintln!(
-                "    inertia.diag=[{:.10e} {:.10e} {:.10e}]",
-                combined_composite_props.inertia.x_axis.x,
-                combined_composite_props.inertia.y_axis.y,
-                combined_composite_props.inertia.z_axis.z
-            );
-            eprintln!(
-                "  orig_parent_cm_struct=[{:.10e} {:.10e} {:.10e}]",
-                orig_parent_cm_struct.x, orig_parent_cm_struct.y, orig_parent_cm_struct.z
-            );
-            eprintln!("=== end ATTACH TRACE ===");
-        }
-
-        // Run the JEOD combine algorithm.
-        let combined = combine_states_at_attach(AttachCombineInputs {
-            parent_composite: parent_composite_pre,
-            parent_mass: parent_pre_composite_props,
-            parent_t_inertial_struct,
-            child_composite,
-            child_mass: subtree_composite_props,
-            combined_mass: combined_composite_props,
-            orig_parent_cm_struct,
-        });
-
-        // The new whole-tree composite state is the integration target —
-        // matches JEOD's `composite_body` post-attach (the source for
-        // `Vel_Rate` per `set_state_source_internal` at the end of
-        // `DynBody::attach_update_properties`).
-        self.bodies[integrated_body_idx].trans = TranslationalState {
-            position: combined.composite_state.trans.position,
-            velocity: combined.composite_state.trans.velocity,
-        };
-        self.bodies[integrated_body_idx].rot = Some(RotationalState {
-            quaternion: combined.composite_state.rot.q_parent_this,
-            ang_vel_body: combined.composite_state.rot.ang_vel_this,
-        });
-        self.bodies[integrated_body_idx].mass = Some(combined_composite_props);
-
-        // Combine succeeded — only now remove the subtree's detached
-        // entry. If any earlier step panicked (missing mass points,
-        // combine preconditions, etc.) the entry survives so callers
-        // can retry or inspect rather than silently losing state.
-        self.detached_subtrees.remove(&subtree_root_id);
-
-        if std::env::var("APOLLO_TRACE").is_ok() {
-            eprintln!(
-                "  COMBINE OUTPUT: pos=[{:.4e} {:.4e} {:.4e}] ω_body=[{:.6e} {:.6e} {:.6e}]",
-                combined.composite_state.trans.position.x,
-                combined.composite_state.trans.position.y,
-                combined.composite_state.trans.position.z,
-                combined.composite_state.rot.ang_vel_this.x,
-                combined.composite_state.rot.ang_vel_this.y,
-                combined.composite_state.rot.ang_vel_this.z
-            );
-        }
-    }
-
-    /// Advance every entry in [`Simulation::detached_subtrees`] by `dt`
-    /// seconds. Each subtree propagates ballistically — no gravity, no
-    /// torque — matching JEOD's behavior for tree roots whose
-    /// `grav_interaction.controls` is empty (which is the case for
-    /// every non-LES vehicle in `SIM_Apollo`).
-    pub fn step_detached_subtrees(&mut self, dt: f64) {
-        for state in self.detached_subtrees.values_mut() {
-            state.step_ballistic(dt);
-        }
-    }
-
-    /// Number of bodies in the simulation.
-    pub fn num_bodies(&self) -> usize {
-        self.bodies.len()
-    }
-
-    /// Set the integration timestep (must be positive).
-    ///
-    /// For JEOD-style time reversal, use `sim.time.time_scale_factor = -1.0`
-    /// instead of negative dt. This keeps `simtime` monotonically increasing
-    /// while reversing dynamic time (TAI, TDB, etc.) and integration direction.
-    ///
-    /// # Panics
-    /// Panics if `dt` is not finite or not positive.
-    pub fn set_dt(&mut self, dt: f64) {
-        assert!(
-            dt.is_finite() && dt > 0.0,
-            "dt must be finite and > 0, got {dt}"
-        );
-        self.dt = dt;
-    }
-
-    /// Current simulation elapsed time in seconds.
-    pub fn elapsed(&self) -> f64 {
-        self.time.simtime
-    }
-}
+// `Simulation` and its supporting public types live in the `simulation`
+// submodule (issue #253). Re-exported here for API stability.
+pub use simulation::{ContactPairConfig, DetachedSubtreeState, Simulation, VehicleOutput};
