+++ crates/jeod_runner/tests/tier3_sim_integ_analytical.rs
@@ -10,11 +10,12 @@
 //! for position and velocity error over a one-orbit propagation interval.
 
 use glam::DVec3;
-use jeod_runner::{GravitySourceEntry, RotationModel, Simulation, VehicleConfig};
+use jeod_runner::{RotationModel, Simulation};
 use jeod_sim::{
     GaussJacksonConfig, GravityControl, GravityControls, GravityModel, GravitySource,
     IntegratorType, SimulationTime, TranslationalState,
 };
+use jeod_sim::{GravitySourceEntry, VehicleConfig};
 
 /// Earth gravitational parameter (m^3/s^2) from JEOD earth_GGM05C.
 const MU_EARTH: f64 = jeod_sim::EARTH.shape.mu;
