+++ crates/jeod_runner/tests/tier3_sim_lsode.rs
@@ -51,11 +51,12 @@
 use jeod_test_data::tier3_csv::test_data_path;
 
 use glam::DVec3;
-use jeod_runner::{GravitySourceEntry, RotationModel, Simulation, VehicleConfig};
+use jeod_runner::{RotationModel, Simulation};
 use jeod_sim::{
     GravityControl, GravityControls, GravityModel, GravitySource, IntegratorType, SimulationTime,
     TranslationalState,
 };
+use jeod_sim::{GravitySourceEntry, VehicleConfig};
 use jeod_test_data::crossval::{CrossvalReport, StateLog};
 use std::path::Path;
