+++ crates/jeod_runner/tests/tier3_sim_lvlh_extended.rs
@@ -15,10 +15,9 @@
 //! No Docker reference data required.
 
 use glam::DVec3;
-use jeod_runner::{
-    DerivedStateConfig, GravitySourceEntry, RotationModel, Simulation, VehicleConfig,
-};
+use jeod_runner::{RotationModel, Simulation};
 use jeod_sim::recipes::helpers::state_helpers::max_mat_diff;
+use jeod_sim::{DerivedStateConfig, GravitySourceEntry, VehicleConfig};
 use jeod_sim::{
     GravityControl, GravityControls, GravityModel, GravitySource, SimulationTime,
     TranslationalState,
