+++ crates/jeod_runner/tests/tier3_sim_lvlh_relative.rs
@@ -7,7 +7,8 @@
 use jeod_test_data::tier3_csv::test_data_path;
 
 use glam::DVec3;
-use jeod_runner::{Simulation, VehicleConfig};
+use jeod_runner::Simulation;
+use jeod_sim::VehicleConfig;
 use jeod_sim::{compute_lvlh_relative_state, SimulationTime, TranslationalState};
 
 struct LvlhRelRecord {
