+++ crates/jeod_runner/tests/tier3_sim_mercury.rs
@@ -12,11 +12,12 @@
 use jeod_test_data::tier3_csv::test_data_path;
 
 use glam::DVec3;
-use jeod_runner::{GravitySourceEntry, Simulation, VehicleConfig};
+use jeod_runner::Simulation;
 use jeod_sim::{
     GravityControl, GravityControls, GravityModel, GravitySource, SimulationTime,
     TranslationalState,
 };
+use jeod_sim::{GravitySourceEntry, VehicleConfig};
 
 fn load_mu_sun() -> f64 {
     jeod_test_data::gravity_fixtures::load_sun_spherical_mu()
