+++ crates/jeod_runner/tests/tier3_sim_orbelem_comprehensive.rs
@@ -8,9 +8,8 @@ use jeod_sim::recipes::helpers::state_helpers::angle_diff;
 use jeod_test_data::tier3_csv::test_data_path;
 
 use glam::DVec3;
-use jeod_runner::{
-    DerivedStateConfig, GravitySourceEntry, RotationModel, Simulation, VehicleConfig,
-};
+use jeod_runner::{RotationModel, Simulation};
+use jeod_sim::{DerivedStateConfig, GravitySourceEntry, VehicleConfig};
 use jeod_sim::{
     GravityControl, GravityControls, GravityModel, GravitySource, SimulationTime,
     TranslationalState,
