+++ crates/jeod_sim/src/vehicle_builder.rs
@@ -9,7 +9,7 @@
 //! the same descriptor that `SimulationBuilder::add_body` accepts and
 //! that `Simulation::body` carries through the runner's step loop.
 //! Phase 6 of #101 consolidated the runtime fluent
-//! `jeod_runner::VehicleBuilder` into this typestate builder, so every
+//! `jeod_sim::VehicleBuilder` into this typestate builder, so every
 //! method that builder offered is available here in the `Ready` impl.
 //!
 //! # Compile-time gating
